% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical.R
\name{Categorical}
\alias{Categorical}
\title{Create a Categorical distribution}
\usage{
Categorical(outcomes = numeric(), p = NULL)
}
\arguments{
\item{outcomes}{A vector specifying the elements in the sample
space. Can be numeric, factor, character, or logical.}

\item{p}{A vector of success probabilities for each outcome.
Each element of \code{p} can be any positive value -- the vector gets
normalized internally. Defaults to \code{NULL}, in which case the
distribution is assumed to be uniform.}
}
\value{
A \code{Categorical} object.
}
\description{
Create a Categorical distribution
}
\examples{

set.seed(27)

X <- Categorical(1:3, p = c(0.4, 0.1, 0.5))
X

Y <- Categorical(LETTERS[1:4])
Y

random(X, 10)
random(Y, 10)

pdf(X, 1)
log_pdf(X, 1)

cdf(X, 1)
quantile(X, 0.5)

# cdfs are only defined for numeric sample spaces. this errors!
# cdf(Y, "a")

# same for quantiles. this also errors!
# quantile(Y, 0.7)
}
\seealso{
Other discrete distributions:
\code{\link[=Bernoulli]{Bernoulli()}},
\code{\link[=Binomial]{Binomial()}},
\code{\link[=Geometric]{Geometric()}},
\code{\link[=HurdleNegativeBinomial]{HurdleNegativeBinomial()}},
\code{\link[=HurdlePoisson]{HurdlePoisson()}},
\code{\link[=HyperGeometric]{HyperGeometric()}},
\code{\link[=Multinomial]{Multinomial()}},
\code{\link[=NegativeBinomial]{NegativeBinomial()}},
\code{\link[=Poisson]{Poisson()}},
\code{\link[=PoissonBinomial]{PoissonBinomial()}},
\code{\link[=ZINegativeBinomial]{ZINegativeBinomial()}},
\code{\link[=ZIPoisson]{ZIPoisson()}},
\code{\link[=ZTNegativeBinomial]{ZTNegativeBinomial()}},
\code{\link[=ZTPoisson]{ZTPoisson()}}
}
\concept{discrete distributions}
