% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{cdf.bernoulli}
\alias{cdf.bernoulli}
\title{Evaluate the cumulative distribution function of a bernoulli distribution}
\usage{
\method{cdf}{bernoulli}(d, x, ...)
}
\arguments{
\item{d}{A `bernoulli` object created by a call to [bernoulli()].}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution `d`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of `x`.
}
\description{
Evaluate the cumulative distribution function of a bernoulli distribution
}
\examples{

b <- bernoulli(0.7)
b

random(b, 10)
pdf(b, 1)
cdf(b, 0)
quantile(b, 0.7)

# TODO: looks like I don't quite have quantiles right since the
# inverses don't hold like I'd want

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
