% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{cdf.beta}
\alias{cdf.beta}
\title{Evaluate the cumulative distribution function of a beta distribution}
\usage{
\method{cdf}{beta}(d, x, ...)
}
\arguments{
\item{d}{A \code{beta} object created by a call to \code{\link[=beta]{beta()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a beta distribution
}
\examples{

b <- beta(1, 2)
b

random(b, 10)
pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
