% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{cdf.binomial}
\alias{cdf.binomial}
\title{Evaluate the cumulative distribution function of a binomial distribution}
\usage{
\method{cdf}{binomial}(d, x, ...)
}
\arguments{
\item{d}{A \code{binomial} object created by a call to \code{\link[=binomial]{binomial()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a binomial distribution
}
\examples{

b <- binomial(10, 0.2)
b

random(b, 10)
pdf(b, 2L)
cdf(b, 4L)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 7))

}
