% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{cdf.cauchy}
\alias{cdf.cauchy}
\title{Evaluate the cumulative distribution function of a cauchy distribution}
\usage{
\method{cdf}{cauchy}(d, x, ...)
}
\arguments{
\item{d}{A \code{cauchy} object created by a call to \code{\link[=cauchy]{cauchy()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a cauchy distribution
}
\examples{

c <- cauchy(10, 0.2)
c

random(c, 10)
pdf(c, 2)
log_pdf(c, 2)
cdf(c, 2)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
