% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{cdf.Erlang}
\alias{cdf.Erlang}
\title{Evaluate the cumulative distribution function of an Erlang distribution}
\usage{
\method{cdf}{Erlang}(d, x, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{Erlang} object created by a call to \code{\link[=Erlang]{Erlang()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Arguments to be passed to \code{\link[stats]{pgamma}}.
Unevaluated arguments will generate a warning to catch mispellings or other
possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of an Erlang distribution
}
\examples{

set.seed(27)

X <- Erlang(5, 2)
X

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 7))
}
