% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_f.R
\name{cdf.fisher_f}
\alias{cdf.fisher_f}
\title{Evaluate the cumulative distribution function of an F distribution}
\usage{
\method{cdf}{fisher_f}(d, x, ...)
}
\arguments{
\item{d}{A \code{fisher_f} object created by a call to \code{\link[=fisher_f]{fisher_f()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of an F distribution
}
\examples{

f <- fisher_f(5, 10, 0.2)
f

random(f, 10)
pdf(f, 2)
cdf(f, 4)
quantile(f, 0.7)

cdf(f, quantile(f, 0.7))
quantile(f, cdf(f, 7))

}
