% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.R
\name{cdf.geometric}
\alias{cdf.geometric}
\title{Evaluate the cumulative distribution function of a geometric distribution}
\usage{
\method{cdf}{geometric}(d, x, ...)
}
\arguments{
\item{d}{A \code{geometric} object created by a call to \code{\link[=geometric]{geometric()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a geometric distribution
}
\examples{

g <- geometric(0.3)
g

random(g, 10)
pdf(g, 2)
log_pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

}
\seealso{
Other geometric distribution: \code{\link{pdf.geometric}},
  \code{\link{quantile.geometric}},
  \code{\link{random.geometric}}
}
\concept{geometric distribution}
