% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gumbel.R
\name{cdf.Gumbel}
\alias{cdf.Gumbel}
\title{Evaluate the cumulative distribution function of a Gumbel distribution}
\usage{
\method{cdf}{Gumbel}(d, x, drop = TRUE, ...)
}
\arguments{
\item{d}{A \code{Gumbel} object created by a call to \code{\link[=Gumbel]{Gumbel()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Arguments to be passed to \code{\link[revdbayes]{pgev}}.
Unevaluated arguments will generate a warning to catch mispellings or other
possible errors.}
}
\value{
In case of a single distribution object, a numeric
vector of cumulative probabilities of length \code{x} (if \code{drop = TRUE}, default)
or a \code{data.frame} with \code{n} columns. In case of a vectorized distribution
object, either a matrix (if \code{drop = TRUE}, default) or a \code{data.frame}
with \code{n} columns, containing all possible combinations.
}
\description{
Evaluate the cumulative distribution function of a Gumbel distribution
}
\examples{

set.seed(27)

X <- Gumbel(1, 2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
