% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{cdf.negative_binomial}
\alias{cdf.negative_binomial}
\title{Evaluate the cumulative distribution function of a negative binomial distribution}
\usage{
\method{cdf}{negative_binomial}(d, x, ...)
}
\arguments{
\item{d}{A \code{negative_binomial} object created by a call to
\code{\link[=negative_binomial]{negative_binomial()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a negative binomial distribution
}
\examples{

nb <- negative_binomial(10, 0.3)
nb

random(nb, 10)
pdf(nb, 2)
log_pdf(nb, 2)
cdf(nb, 4)
quantile(nb, 0.7)

}
\seealso{
Other negative_binomial distribution: \code{\link{pdf.negative_binomial}},
  \code{\link{quantile.negative_binomial}},
  \code{\link{random.negative_binomial}}
}
\concept{negative_binomial distribution}
