% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{cdf.poisson}
\alias{cdf.poisson}
\title{Evaluate the cumulative distribution function of a poisson distribution}
\usage{
\method{cdf}{poisson}(d, x, ...)
}
\arguments{
\item{d}{A \code{poisson} object created by a call to \code{\link[=poisson]{poisson()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a poisson distribution
}
\examples{

p <- poisson(2)
p

random(p, 10)
pdf(p, 2)
cdf(p, 4)
quantile(p, 0.7)

cdf(p, quantile(p, 0.7))
quantile(p, cdf(p, 7))

}
