% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t.R
\name{cdf.students_t}
\alias{cdf.students_t}
\title{Evaluate the cumulative distribution function of a students_t distribution}
\usage{
\method{cdf}{students_t}(d, x, ...)
}
\arguments{
\item{d}{A \code{students_t} object created by a call to \code{\link[=students_t]{students_t()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a students_t distribution
}
\examples{

# TODO

}
\seealso{
Other students_t distribution: \code{\link{pdf.students_t}},
  \code{\link{quantile.students_t}},
  \code{\link{random.students_t}}
}
\concept{students_t distribution}
