% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentsT.R
\name{cdf.StudentsT}
\alias{cdf.StudentsT}
\title{Evaluate the cumulative distribution function of a StudentsT distribution}
\usage{
\method{cdf}{StudentsT}(d, x, ...)
}
\arguments{
\item{d}{A \code{StudentsT} object created by a call to \code{\link[=StudentsT]{StudentsT()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a StudentsT distribution
}
\examples{

# TODO

}
\seealso{
Other StudentsT distribution: \code{\link{pdf.StudentsT}},
  \code{\link{quantile.StudentsT}},
  \code{\link{random.StudentsT}}
}
\concept{StudentsT distribution}
