% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.R
\name{cdf.uniform}
\alias{cdf.uniform}
\title{Evaluate the cumulative distribution function of a continuous uniform distribution}
\usage{
\method{cdf}{uniform}(d, x, ...)
}
\arguments{
\item{d}{A \code{uniform} object created by a call to \code{\link[=uniform]{uniform()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the cumulative distribution function of a continuous uniform distribution
}
\examples{

b <- uniform(1, 2)
b

random(b, 10)
pdf(b, 0.7)
log_pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
