% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroAdjustedNegbin.R
\name{cdf.ZeroAdjustedNegbin}
\alias{cdf.ZeroAdjustedNegbin}
\title{Evaluate the cumulative distribution function of a Zero-Adjusted Negative Binomial distribution}
\usage{
\method{cdf}{ZeroAdjustedNegbin}(d, x, drop = TRUE, elementwise = NULL, cores = NULL, ...)
}
\arguments{
\item{d}{A \code{ZeroAdjustedNegbin} object created by a call to \code{\link[=ZeroAdjustedNegbin]{ZeroAdjustedNegbin()}}.}

\item{x}{A vector of elements whose cumulative probabilities you would like to
determine given the distribution \code{d}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{elementwise}{logical. Should each distribution in \code{d} be evaluated
at all elements of \code{x} (\code{elementwise = FALSE}, yielding a matrix)?
Or, if \code{d} and \code{x} have the same length, should the evaluation be
done element by element (\code{elementwise = TRUE}, yielding a vector)? The
default of \code{NULL} means that \code{elementwise = TRUE} is used if the
lengths match and otherwise \code{elementwise = FALSE} is used.}

\item{cores}{\code{NULL} or integer. Number of cores/threads to use when calling \code{\link[=pZeroAdjustedNegbin]{pZeroAdjustedNegbin()}}.}

\item{...}{Arguments to be passed to \code{\link[stats]{pnbinom}}.
Unevaluated arguments will generate a warning to catch mispellings or other
possible errors.}
}
\value{
In case of a single distribution object, either a numeric
vector of length \code{length(x)} (if \code{drop = TRUE}, default) or a \code{matrix} with
\code{length(x)} columns (if \code{drop = FALSE}). In case of a vectorized distribution
object, a matrix with \code{length(x)} columns containing all possible combinations.
}
\description{
Evaluate the cumulative distribution function of a Zero-Adjusted Negative Binomial distribution
}
