% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{chi_square}
\alias{chi_square}
\title{Create a chi square distribution}
\usage{
chi_square(df)
}
\arguments{
\item{df}{Degrees of freedom. Must be positive.}
}
\value{
A \code{chi_square} object.
}
\description{
Create a chi square distribution
}
\examples{

c <- chi_square(5)
c

random(c, 10)
pdf(c, 2)
cdf(c, 4)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{exponential}},
  \code{\link{fisher_f}}, \code{\link{gamma}},
  \code{\link{normal}}, \code{\link{poisson}}
}
\concept{continuous distributions}
