% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{Custom}
\alias{Custom}
\title{Create a custom discrete distribution}
\usage{
Custom(ss = NULL, probs = NULL)
}
\arguments{
\item{ss}{A vector specifying the sample space.}

\item{probs}{A vector of same length as \code{ss} giving the probabilities of each outcome. Must sum to 1.
Defaults to \code{rep(1/length(ss), length(ss))}}
}
\value{
A \code{Custom} object.
}
\description{
Create a custom discrete distribution
}
\examples{

X <- Custom(ss = c(1,6,2), probs = c(0.4, 0.1, 0.5))
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))

}
\seealso{
Other discrete distributions: \code{\link{Bernoulli}},
  \code{\link{Binomial}}, \code{\link{Geometric}},
  \code{\link{HyperGeometric}}, \code{\link{Multinomial}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}}
}
\concept{discrete distributions}
