% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv2.R
\name{deriv2}
\alias{deriv2}
\alias{deriv2.SinhArcsinh}
\alias{deriv2.Normal}
\title{Experimental deriv2 method}
\usage{
deriv2(d, ...)

\method{deriv2}{SinhArcsinh}(
  d,
  x,
  which.score = NULL,
  which.hessian = NULL,
  drop = TRUE,
  expected = FALSE,
  cores = 1L,
  ...
)

\method{deriv2}{Normal}(
  d,
  x,
  which.score = NULL,
  which.hessian = NULL,
  drop = TRUE,
  expected = FALSE,
  cores = 1L,
  ...
)
}
\arguments{
\item{d}{An object, typically a distribution object, e.g., as created by
\code{\link[=Normal]{Normal()}} or \code{\link[=Binomial]{Binomial()}}.}

\item{...}{Arguments passed to methods.}

\item{x}{A vector of elements whose score/Hessian should be determined given the
distribution \code{d}. Either \code{d} and \code{x} need to have the same length or length 1.}

\item{which.score}{character, \code{NULL} (default), or \code{NA}. Character labels
for the derivatives to be included in the score. Possible values are
(combinations of) the parameter names (e.g., \code{"mu"} and/or \code{"sigma"}).
By default (if \code{which.score = NULL}) all elements of the score
should be computed. If set \code{NA}, no scores are returned.}

\item{which.hessian}{character, \code{NULL} (default), or \code{NA}. Character labels
for the derivatives and cross-derivatives are available
(e.g., \code{mu}, \code{sigma}, \code{mu:sigma} and \code{sigma:mu})
By default (if \code{which.hessian = NULL}) all elements of the score or Hessian
should be computed. If set \code{NA}, no hessians are returned.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{expected}{logical. Should the expected Hessian be computed? If \code{FALSE} the
observed Hessian is computed. Some methods might only support only one or the
the other option and defaults might differ.}

\item{cores}{integer, number of cores to be used if OpenMP is available.
defaults to \code{1L}.}
}
\value{
Named list with the requested scores an hessian elements.
If nothing is requested (i.e., \code{which.score} or \code{which.hessian} are
set to \code{NA}) the list element contains \code{NULL}, else a vector
or matrix with the corresponding scores/hessian (see \code{drop}).
}
\description{
Experimental deriv2 method
}
\examples{
X <- SinhArcsinh(mu    = as.double(1:3),
                 sigma = c(0.5, 1.5, 1.0),
                 nu    = c(0.9, 1.0, 1.1),
                 tau   = c(1.0, 0.8, 0.6))
names(X) <- LETTERS[1:3]
X

x <- c(-5, 10, 0.3)

## Calculate all scores and full hessian
deriv2(X, x)

## Only score/only hessian
deriv2(X, x, which.hessian = NA)
deriv2(X, x, which.score = NA)

## Only dl/dmu and d2l/dmu^2
deriv2(X, x, which.score = "mu", which.hessian = "mu")
deriv2(X, x, which.score = "mu", which.hessian = "mu", drop = FALSE)

## Custom set of derivatives
deriv2(X, x, which.score = c("nu", "tau"),
       which.hessian = c("nu:tau", "tau:nu", "sigma"), drop = FALSE)

}
