% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-package.R
\docType{package}
\name{distributions-package}
\alias{distributions}
\alias{distributions-package}
\title{distributions: Probability Distributions as S3 Objects}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/distributions}
  \item Report bugs at \url{https://github.com/alexpghayes/distributions/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (0000-0002-4985-5160)

}
\keyword{internal}
