% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroAdjustedNegbin.R
\name{dZeroAdjustedNegbin}
\alias{dZeroAdjustedNegbin}
\alias{pZeroAdjustedNegbin}
\alias{qZeroAdjustedNegbin}
\alias{rZeroAdjustedNegbin}
\title{The Zero-Adjusted Negative Binomial distribution}
\usage{
dZeroAdjustedNegbin(x, mu = 1, sigma = 1, nu = 0.3, log = FALSE, cores = NULL)

pZeroAdjustedNegbin(
  q,
  mu = 1,
  sigma = 1,
  nu = 0.3,
  lower.tail = TRUE,
  log.p = FALSE,
  cores = NULL
)

qZeroAdjustedNegbin(
  p,
  mu = 1,
  sigma = 1,
  nu = 0.3,
  lower.tail = TRUE,
  log.p = FALSE,
  cores = NULL
)

rZeroAdjustedNegbin(n, mu = 1, sigma = 1, nu = 0.3, cores = NULL)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{mu, sigma, nu}{vector of (positive) parameters.}

\item{log, log.p}{logical indicating whether probabilities p are given as log(p).}

\item{cores}{\code{NULL} or positive integer. If not \code{NULL} we use the C code with \code{cores} threads.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical indicating whether probabilities are \eqn{P[X <= x]} (TRUE) or \eqn{P[X > x]} (FALSE).}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\description{
Density, distribution function, quantile function, and random
generation for the zero-adjusted negative binomial distribution with three
parameters \code{mu}, \code{sigma}, and \code{nu}.
}
\details{
All functions follow the usual conventions of d/p/q/r functions in base R.
}
\examples{
## PMF
x <- 0:10
p <- dZeroAdjustedNegbin(x, mu = 2, sigma = 0.5, nu = 0.3)
plot(x, p, type = "h", lwd = 2)

## CDF and quantile roundtrip
pZeroAdjustedNegbin(qZeroAdjustedNegbin(0.5, mu = 2, sigma = 0.5, nu = 0.3),
                    mu = 2, sigma = 0.5, nu = 0.3)

}
\concept{ZeroAdjustedNegbin}
\keyword{distribution}
