% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_f.R
\name{fisher_f}
\alias{fisher_f}
\title{Create an F distribution}
\usage{
fisher_f(df1, df2, lambda = 0)
}
\arguments{
\item{df1}{Numerator degrees of freedom. Can be any positive number.}

\item{df2}{Denominator degrees of freedom. Can be any positive number.}

\item{lambda}{Non-centrality parameter. Can be any positive number.
Defaults to \code{0}.}
}
\value{
An \code{f} object.
}
\description{
Create an F distribution
}
\examples{

f <- fisher_f(5, 10, 0.2)
f

random(f, 10)
pdf(f, 2)
log_pdf(f, 2)
cdf(f, 4)
quantile(f, 0.7)

cdf(f, quantile(f, 0.7))
quantile(f, cdf(f, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{chi_square}},
  \code{\link{exponential}}, \code{\link{gamma}},
  \code{\link{normal}}, \code{\link{poisson}},
  \code{\link{students_t}}
}
\concept{continuous distributions}
