% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fit}
\alias{fit}
\title{Fit a distribution to data}
\usage{
fit(d, x, ...)
}
\arguments{
\item{d}{A probability distribution object such as those created by
a call to \code{\link[=bernoulli]{bernoulli()}}, \code{\link[=beta]{beta()}}, or \code{\link[=binomial]{binomial()}}.}

\item{x}{A vector of data to compute the likelihood.}
}
\value{
an object the same class as \code{d} with updated parameters
}
\description{
Approximates an empirical distribution with a theoretical one
}
