% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{fit_mle.bernoulli}
\alias{fit_mle.bernoulli}
\title{Fit a bernoulli distribution to data}
\usage{
\method{fit_mle}{bernoulli}(d, x, ...)
}
\arguments{
\item{d}{A \code{bernoulli} object.}

\item{x}{A vector of zeroes and ones.}
}
\value{
a \code{bernoulli} object
}
\description{
Fit a bernoulli distribution to data
}
