% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{fit_mle.binomial}
\alias{fit_mle.binomial}
\title{Fit a binomial distribution to data}
\usage{
\method{fit_mle}{binomial}(d, x, ...)
}
\arguments{
\item{d}{A \code{binomial} object.}

\item{x}{A vector of zeroes and ones to fit the binomial distribution to.}
}
\value{
a \code{binomial} object
}
\description{
The fit distribution will inherit the same \code{size} parameter as
the \code{binomial} object passed.
}
