% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{fit_mle.Erlang}
\alias{fit_mle.Erlang}
\title{Fit a Erlang distribution to data}
\usage{
\method{fit_mle}{Erlang}(d, x, ...)
}
\arguments{
\item{d}{A \code{Erlang} object created by a call to \code{\link[=Erlang]{Erlang()}}.}

\item{x}{A vector to fit the Erlang distribution to.}

\item{...}{Unused.}
}
\value{
a \code{Erlang} object
}
\description{
Fit a Erlang distribution to data
}
