% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{fit_mle.exponential}
\alias{fit_mle.exponential}
\title{Fit an exponential distribution to data}
\usage{
\method{fit_mle}{exponential}(d, x, ...)
}
\arguments{
\item{d}{An \code{exponential} object created by a call to \code{\link[=exponential]{exponential()}}.}

\item{x}{A vector of data.}
}
\value{
An \code{exponential} object.
}
\description{
Fit an exponential distribution to data
}
\concept{exponential distribution}
