% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{fit_mle.gamma}
\alias{fit_mle.gamma}
\title{Fit a gamma distribution to data}
\usage{
\method{fit_mle}{gamma}(d, x, ...)
}
\arguments{
\item{d}{A \code{gamma} object created by a call to \code{\link[=gamma]{gamma()}}.}

\item{x}{A vector to fit the gamma distribution to.}
}
\value{
a \code{gamma} object
}
\description{
Fit a gamma distribution to data
}
