% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric.R
\name{fit_mle.Geometric}
\alias{fit_mle.Geometric}
\title{Fit a Geometric distribution to data}
\usage{
\method{fit_mle}{Geometric}(d, x, ...)
}
\arguments{
\item{d}{A \code{Geometric} object.}

\item{x}{A vector of zeroes and ones.}

\item{...}{Unused.}
}
\value{
a \code{Geometric} object
}
\description{
Fit a Geometric distribution to data
}
