% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace.R
\name{fit_mle.Laplace}
\alias{fit_mle.Laplace}
\title{Fit a Laplace distribution to data}
\usage{
\method{fit_mle}{Laplace}(d, x, ...)
}
\arguments{
\item{d}{A \code{Laplace} object created by a call to \code{\link[=Laplace]{Laplace()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A \code{Laplace} object.
}
\description{
Fit a Laplace distribution to data
}
\concept{Laplace distribution}
