% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{fit_mle.Pareto}
\alias{fit_mle.Pareto}
\title{Fit a Pareto distribution to data}
\usage{
\method{fit_mle}{Pareto}(d, x, ...)
}
\arguments{
\item{d}{A \code{Pareto} object created by a call to \code{\link[=Pareto]{Pareto()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A \code{Pareto} object.
}
\description{
Fit a Pareto distribution to data
}
\concept{Pareto distribution}
