% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{fit_mle.poisson}
\alias{fit_mle.poisson}
\title{Fit an poisson distribution to data}
\usage{
\method{fit_mle}{poisson}(d, x, ...)
}
\arguments{
\item{d}{An \code{poisson} object created by a call to \code{\link[=poisson]{poisson()}}.}

\item{x}{A vector of data.}
}
\value{
An \code{poisson} object.
}
\description{
Fit an poisson distribution to data
}
\concept{poisson distribution}
