% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_prob}
\alias{geom_prob}
\title{Show probabilities as area under curve}
\usage{
geom_prob(from = -Inf, to = Inf, digits = 3, annotate = FALSE, ...)
}
\arguments{
\item{from}{Left end-point of interval}

\item{to}{right end-point of interval}

\item{digits}{number of digits to include when printing probability}

\item{annotate}{logical. If FALSE (default), the probability is not shown on plot.}
}
\description{
A function that let's us draw area under curve, and with option to annotate plot
with P(from < X < to).
}
\examples{

X <- Normal()

#plot_pdf(X) + geom_prob(to = -0.645)
#plot_pdf(X) + geom_prob(from = -0.645, to = 0.1)

}
