% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-hessian.R
\name{get_deriv_params}
\alias{get_deriv_params}
\title{Get Parameter Names for Derivatives}
\usage{
get_deriv_params(p, which = NULL, expand = FALSE)
}
\arguments{
\item{p}{character vector with the names of the parameters
of the distribution.}

\item{which}{\code{NULL} or character vector. If not \code{NULL}
(used for input sanity check) match.arg is used
to return parameter names requested by the users.}

\item{expand}{logical, defaults to \code{FALSE}. If set \code{TRUE},
cross-derivative parameters are calculated.
The score method uses \code{expand = FALSE}, the hessian
uses \code{expand = TRUE}.}
}
\value{
Named vector. If \code{which = NULL} a named vector
with all potential parameters is returned, including
parameters for cross-derivatives if \code{expand = TRUE}.
}
\description{
Get Parameter Names for Derivatives
}
