% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasS3method}
\alias{hasS3method}
\title{Check available support for S3 method}
\usage{
hasS3method(method, classes)
}
\arguments{
\item{method}{character, name of the method (e.g., \code{"is_continuous"}, \code{"print"}, ...)}

\item{classes}{character vector of length > 0, classes to check.}
}
\value{
Returns \code{TRUE} if the method exists for one of the given classes, else \code{FALSE}.
}
\description{
Evaluates whether or not there is support for a given S3 method for specific
objects.
}
\keyword{internal}
