% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_distribution}
\alias{is_distribution}
\title{Is object a distribution?}
\usage{
is_distribution(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
\code{is_distribution} tests if \code{x} is an distribution.
It simply checks that the object has a \code{"distribution"}
subclass.
}
\examples{
n <- normal()
is_distribution(n)
}
