% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace.R
\name{Laplace}
\alias{Laplace}
\title{Create a Laplace distribution}
\usage{
Laplace(theta = 0, phi = 1)
}
\arguments{
\item{theta}{The location parameter \eqn{\theta}.
\code{theta} can be any real number.  Defaults to \code{0}.}

\item{phi}{The scale parameter \eqn{\phi}.
\code{phi} can be any value strictly greater than zero.  Defaults to \code{1}.}
}
\value{
A \code{Laplace} object.
}
\description{
The Laplace distribution is related to the exponential distribution (see
\code{\link{Exponential}}).  If \eqn{X} has a Laplace distribution with location
parameter \eqn{\theta} and scale parameter \eqn{\phi} then
\eqn{| X - \theta |} has an exponential distribution with rate
parameter \eqn{1 / \phi}.  If \eqn{\theta} is zero then a Laplace
distribution may be viewed as a composite of two identical exponential
distributions placed back-to-back.  Therefore, it may be referred to as a
\emph{double exponential}, \emph{two-tailed exponential} or \emph{bilateral exponential}
distribution.
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions3}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a Laplace random variable with location
parameter \code{theta} = \eqn{\theta} and scale parameter \code{phi} = \eqn{\phi}.

\strong{Support}: \eqn{R}, the set of all real numbers

\strong{Mean, median, mode}: \eqn{\theta}

\strong{Variance}: \eqn{2 \phi^2}

\strong{Probability density function (p.d.f)}:

\deqn{
    f(x) = \frac{1}{2} \phi^{-1} \exp(- |x - \theta)| / \phi)
  }{
    f(x) = (1 / 2 \phi) exp(- |x - \theta)| / \phi)
  }

\strong{Cumulative distribution function (c.d.f)}:

\deqn{
    F(x) = \frac{1}{2} + \frac{1}{2} {\rm ~sgn}(x - \theta)%
     [1 - \exp(-|x - \theta| / \phi)]
  }{
    F(x) = 1/2 + (1/2) sgn(x - \theta) [1 - exp(-|x - \theta| / \phi)]
  }
where sgn() is the sign function: sgn(\eqn{x}) is equal to \eqn{-1} if
\eqn{x} is negative, \eqn{0} if \eqn{x} is zero and \eqn{1} if \eqn{x} is
positive.

\strong{Moment generating function (m.g.f)}:

\deqn{
    E(e^{tX}) = \exp(\theta t) / (1 - \phi^2 t^2)
  }{
    E(e^(tX)) = exp(\theta t) / (1 - \phi^2 t^2)
  }
for \eqn{|t| < 1 / \phi}.
}
\examples{

set.seed(27)

X <- Laplace(1, 2)
X

random(X, 10)

pdf(X, c(0.3, 1.7))
log_pdf(X, c(0.3, 1.7))

cdf(X, c(0.3, 1, 1.7))
quantile(X, c(0.25, 0.5, 0.75))

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))

x <- random(X, 100)
fit_mle(X, x)
}
\seealso{
Other continuous distributions: \code{\link{Beta}},
  \code{\link{Cauchy}}, \code{\link{ChiSquare}},
  \code{\link{Exponential}}, \code{\link{FisherF}},
  \code{\link{Gamma}}, \code{\link{LogNormal}},
  \code{\link{Logistic}}, \code{\link{Normal}},
  \code{\link{StudentsT}}, \code{\link{Tukey}},
  \code{\link{Uniform}}, \code{\link{Weibull}}
}
\concept{continuous distributions}
