% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{likelihood}
\alias{likelihood}
\title{Compute the likelihood of a probability distribution given data}
\usage{
likelihood(d, x, ...)
}
\arguments{
\item{d}{A probability distribution object such as those created by
a call to \code{\link[=bernoulli]{bernoulli()}}, \code{\link[=beta]{beta()}}, or \code{\link[=binomial]{binomial()}}.}

\item{x}{A vector of data to compute the likelihood.}
}
\value{
the likelihood
}
\description{
Compute the likelihood of a probability distribution given data
}
