% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal.R
\name{log_normal}
\alias{log_normal}
\title{Create a log_normal distribution}
\usage{
log_normal(log_mu = 0, log_sigma = 1)
}
\arguments{
\item{log_mu}{TODO explain what the hell this is}

\item{log_sigma}{TODO explain what the hell this is}
}
\value{
A \code{log_normal} object.
}
\description{
Create a log_normal distribution
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a log_normal random variable with
success probability \code{p} = \eqn{p}.

TODO: multiple parameterizations BLEH

\strong{Support}: TODO

\strong{Mean}: TODO

\strong{Variance}: TODO

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:

TODO
}
\examples{

ln <- log_normal(0.3, 2)
ln

random(ln, 10)
pdf(ln, 2)
log_pdf(ln, 2)
cdf(ln, 4)
quantile(ln, 0.7)

}
\seealso{
Other discrete distributions: \code{\link{bernoulli}},
  \code{\link{binomial}}, \code{\link{geometric}},
  \code{\link{negative_binomial}}
}
\concept{discrete distributions}
