% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_positive_integer}
\alias{make_positive_integer}
\title{Helper Function to Make a Positive Integer}
\usage{
make_positive_integer(n)
}
\arguments{
\item{n}{numeric. Typically number of observations for computing random draws. If \code{length(n) > 1}, the length is taken to be the number required (consistent with base R as, e.g., for \code{rnorm()}.}
}
\description{
Helper Function to Make a Positive Integer
}
