% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_support}
\alias{make_support}
\title{Helper Function to Return Support for \code{distributions3} objectfor \code{distributions3} object}
\usage{
make_support(min, max, d, drop = TRUE)
}
\arguments{
\item{min, max}{Numeric vectors. Minima and Maxima supports of a \code{distributions3} object.}

\item{d}{A \code{distributions3} object.}

\item{drop}{logical. Should the result be simplified to an unnamed vector if possible?}
}
\description{
Helper Function to Return Support for \code{distributions3} objectfor \code{distributions3} object
}
