% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherF.R
\name{mean.FisherF}
\alias{mean.FisherF}
\title{The k-th moment of an F(df1, df2) distribution exists and
is finite only when 2k < d2}
\usage{
\method{mean}{FisherF}(d, ...)
}
\description{
The k-th moment of an F(df1, df2) distribution exists and
is finite only when 2k < d2
}
