% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{negative_binomial}
\alias{negative_binomial}
\title{Create a negative_binomial distribution}
\usage{
negative_binomial(size, p = 0.5)
}
\arguments{
\item{size}{The number of trials. Must be an integer greater than or equal
to one. When \code{size = 1L}, the binomial distribution reduces to the
bernoulli distribution. Oftened called \code{n} in textbooks.}

\item{p}{The success probability for a given trial. \code{p} can be any
value in \code{[0, 1]}, and defaults to \code{0.5}.}
}
\value{
A \code{negative_binomial} object.
}
\description{
TODO
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a negative_binomial random variable with
success probability \code{p} = \eqn{p}.

TODO: multiple parameterizations BLEH

\strong{Support}: TODO

\strong{Mean}: TODO

\strong{Variance}: TODO

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:

TODO
}
\examples{

nb <- negative_binomial(10, 0.3)
nb

random(nb, 10)
pdf(nb, 2)
log_pdf(nb, 2)
cdf(nb, 4)
quantile(nb, 0.7)

}
\seealso{
Other discrete distributions: \code{\link{Bernoulli}},
  \code{\link{Binomial}}, \code{\link{geometric}},
  \code{\link{log_normal}}, \code{\link{poisson}}
}
\concept{discrete distributions}
