% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter_names}
\alias{parameter_names}
\title{Return parameter names of a distribution object?}
\usage{
parameter_names(x)
}
\arguments{
\item{x}{An distrubtion object.}
}
\description{
\code{parameter_names} returns the parameter names of a \code{"distribution"} object.
}
\examples{

Z <- Normal()

parameter_names(Z)
}
