% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{Pareto}
\alias{Pareto}
\title{Create a Pareto distribution}
\usage{
Pareto(k = 1, a = 1)
}
\arguments{
\item{k}{The scale parameter. \code{k} can be any positive number.
Defaults to \code{1}.}

\item{a}{The shape parameter. \code{a} can be any positive number.
Defaults to \code{1}.}
}
\value{
A \code{Pareto} object.
}
\description{
The (Type I) Pareto distribution is a power-law distribution originating
from Pareto's Law, which was used to describe the distribution of wealth in
a society.  It is also known as \emph{Pareto distribution of the first kind}.
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions3}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a Pareto random variable with scale
parameter \eqn{k} and scale parameter \eqn{a}.

\strong{Support}: \eqn{[k, \infty)}.

\strong{Mean}: \eqn{a k / (a - 1)}, for \eqn{a > 1}.
Undefined otherwise.

\strong{Median}: \eqn{2 ^ {1 / a} k}{2 ^ (1 / a) k}.

\strong{Mode}: \eqn{k}.

\strong{Variance}: \eqn{a k ^ 2 / (a - 1)^2 (a - 2)}, for \eqn{a > 2}.
Undefined otherwise.

\strong{Probability density function (p.d.f)}:

\deqn{
    f(x) = a k^a / x^{a + 1}
  }{
    f(x) = a k^a / x^(a + 1)
  }
for \eqn{x \geq k}{x >= k}.

\strong{Cumulative distribution function (c.d.f)}:

\deqn{
    F(x) = 1 - (k / x)^a
  }{
    F(x) = 1 - (k / x)^a
  }
for \eqn{x \geq k}{x >= k}.

\strong{Moment generating function (m.g.f)}:

\deqn{
    E(e^{tX}) = a(-kt)^a \Gamma(-a, -kt)
  }{
    E(e^(tX)) = a(-kt)^a \Gamma(-a, -kt)
  }
for \eqn{t < 0}, where \eqn{\Gamma(a, x)} is the (upper) incomplete gamma
function. For \eqn{t = 0}, \eqn{E(e^{tX}) = 1}.  The m.g.f. is undefined
for \eqn{t > 0}.
}
\examples{

set.seed(27)

X <- Pareto(1, 2)
X

random(X, 10)

pdf(X, 1.7)
log_pdf(X, 1.7)

cdf(X, 1.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 1.7))

x <- random(X, 100)
fit_mle(X, x)
}
\seealso{
Other continuous distributions: \code{\link{Beta}},
  \code{\link{Cauchy}}, \code{\link{ChiSquare}},
  \code{\link{Exponential}}, \code{\link{FisherF}},
  \code{\link{Gamma}}, \code{\link{LogNormal}},
  \code{\link{Logistic}}, \code{\link{Normal}},
  \code{\link{StudentsT}}, \code{\link{Tukey}},
  \code{\link{Uniform}}, \code{\link{Weibull}}
}
\concept{continuous distributions}
