% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{pdf.beta}
\alias{pdf.beta}
\title{Evaluate the probability mass function of a beta distribution}
\usage{
\method{pdf}{beta}(d, x)
}
\arguments{
\item{d}{A `beta` object created by a call to [beta()].}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution `d`.}
}
\value{
A vector of probabilities, one for each element of `x`.
}
\description{
Evaluate the probability mass function of a beta distribution
}
\examples{

b <- beta(1, 2)
b

random(b, 10)
pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
