% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{pdf.binomial}
\alias{pdf.binomial}
\title{Evaluate the probability mass function of a binomial distribution}
\usage{
\method{pdf}{binomial}(d, x, ...)
}
\arguments{
\item{d}{A `binomial` object created by a call to [binomial()].}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution `d`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of `x`.
}
\description{
Evaluate the probability mass function of a binomial distribution
}
\examples{

b <- binomial(10, 0.2)
b

random(b, 10)
pdf(b, 2L)
cdf(b, 4L)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 7))

}
