% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{pdf.chi_square}
\alias{pdf.chi_square}
\alias{log_pdf.chi_square}
\title{Evaluate the probability mass function of a chi square distribution}
\usage{
\method{pdf}{chi_square}(d, x, ...)

\method{log_pdf}{chi_square}(d, x, ...)
}
\arguments{
\item{d}{A \code{chi_square} object created by a call to \code{\link[=chi_square]{chi_square()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a chi square distribution
}
\examples{

c <- chi_square(5)
c

random(c, 10)
pdf(c, 2)
log_pdf(c, 2)
cdf(c, 4)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
