% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{pdf.exponential}
\alias{pdf.exponential}
\title{Evaluate the probability mass function of a exponential distribution}
\usage{
\method{pdf}{exponential}(d, x, ...)
}
\arguments{
\item{d}{A \code{exponential} object created by a call to \code{\link[=exponential]{exponential()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a exponential distribution
}
\examples{

e <- exponential(5)
e

random(e, 10)
pdf(e, 2)
cdf(e, 4)
quantile(e, 0.7)

cdf(e, quantile(e, 0.7))
quantile(e, cdf(e, 7))

}
