% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.R
\name{pdf.geometric}
\alias{pdf.geometric}
\title{Evaluate the probability mass function of a geometric distribution}
\usage{
\method{pdf}{geometric}(d, x, ...)
}
\arguments{
\item{d}{A \code{geometric} object created by a call to \code{\link[=geometric]{geometric()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Please see the documentation of \code{\link[=geometric]{geometric()}} for some properties
of the geometric distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

g <- geometric(0.3)
g

random(g, 10)
pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

}
\seealso{
Other geometric distribution: \code{\link{cdf.geometric}},
  \code{\link{quantile.geometric}},
  \code{\link{random.geometric}}
}
\concept{geometric distribution}
