% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace.R
\name{pdf.Laplace}
\alias{pdf.Laplace}
\alias{log_pdf.Laplace}
\title{Evaluate the probability mass function of a Laplace distribution}
\usage{
\method{pdf}{Laplace}(d, x, ...)

\method{log_pdf}{Laplace}(d, x, ...)
}
\arguments{
\item{d}{A \code{Laplace} object created by a call to \code{\link[=Laplace]{Laplace()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a Laplace distribution
}
\examples{

set.seed(27)

X <- Laplace(1, 2)
X

random(X, 10)

pdf(X, c(0.3, 1.7))
log_pdf(X, c(0.3, 1.7))

cdf(X, c(0.3, 1, 1.7))
quantile(X, c(0.25, 0.5, 0.75))

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))

x <- random(X, 100)
fit_mle(X, x)
}
