% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal.R
\name{pdf.log_normal}
\alias{pdf.log_normal}
\alias{log_pdf.log_normal}
\title{Evaluate the probability mass function of a log_normal distribution}
\usage{
\method{pdf}{log_normal}(d, x, ...)

\method{log_pdf}{log_normal}(d, x, ...)
}
\arguments{
\item{d}{A \code{log_normal} object created by a call to \code{\link[=log_normal]{log_normal()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Please see the documentation of \code{\link[=log_normal]{log_normal()}} for some properties
of the log_normal distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

ln <- log_normal(0.3, 2)
ln

random(ln, 10)
pdf(ln, 2)
log_pdf(ln, 2)
cdf(ln, 4)
quantile(ln, 0.7)

}
\seealso{
Other log_normal distribution: \code{\link{cdf.log_normal}},
  \code{\link{quantile.log_normal}},
  \code{\link{random.log_normal}}
}
\concept{log_normal distribution}
