% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{pdf.negative_binomial}
\alias{pdf.negative_binomial}
\alias{log_pdf.negative_binomial}
\title{Evaluate the probability mass function of a negative_binomial distribution}
\usage{
\method{pdf}{negative_binomial}(d, x, ...)

\method{log_pdf}{negative_binomial}(d, x, ...)
}
\arguments{
\item{d}{A \code{negative_binomial} object created by a call to
\code{\link[=negative_binomial]{negative_binomial()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a negative_binomial distribution
}
\examples{

nb <- negative_binomial(10, 0.3)
nb

random(nb, 10)
pdf(nb, 2)
log_pdf(nb, 2)
cdf(nb, 4)
quantile(nb, 0.7)

}
\seealso{
Other negative_binomial distribution: \code{\link{cdf.negative_binomial}},
  \code{\link{quantile.negative_binomial}},
  \code{\link{random.negative_binomial}}
}
\concept{negative_binomial distribution}
