% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{pdf.Pareto}
\alias{pdf.Pareto}
\alias{log_pdf.Pareto}
\title{Evaluate the probability mass function of a Pareto distribution}
\usage{
\method{pdf}{Pareto}(d, x, ...)

\method{log_pdf}{Pareto}(d, x, ...)
}
\arguments{
\item{d}{A \code{Pareto} object created by a call to \code{\link[=Pareto]{Pareto()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a Pareto distribution
}
\examples{

set.seed(27)

X <- Pareto(1, 2)
X

random(X, 10)

pdf(X, 1.7)
log_pdf(X, 1.7)

cdf(X, 1.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 1.7))

x <- random(X, 100)
fit_mle(X, x)
}
