% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{pdf.SHASH}
\alias{pdf.SHASH}
\alias{log_pdf.SHASH}
\title{Evaluate the probability mass function of a SHASH distribution}
\usage{
\method{pdf}{SHASH}(d, x, drop = TRUE, elementwise = NULL, cores = NULL, ...)

\method{log_pdf}{SHASH}(d, x, drop = TRUE, elementwise = NULL, cores = NULL, ...)
}
\arguments{
\item{d}{A \code{SHASH} object created by a call to \code{\link[=SHASH]{SHASH()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{elementwise}{logical. Should each distribution in \code{d} be evaluated
at all elements of \code{x} (\code{elementwise = FALSE}, yielding a matrix)?
Or, if \code{d} and \code{x} have the same length, should the evaluation be
done element by element (\code{elementwise = TRUE}, yielding a vector)? The
default of \code{NULL} means that \code{elementwise = TRUE} is used if the
lengths match and otherwise \code{elementwise = FALSE} is used.}

\item{cores}{\code{NULL} or positive integer. TODO(R): Just a development option.
If not \code{NULL} we use the C code with \code{cores} threads.}

\item{...}{Arguments to be passed to \code{\link[stats]{dnorm}}.
Unevaluated arguments will generate a warning to catch mispellings or other
possible errors.}
}
\value{
In case of a single distribution object, either a numeric
vector of length \code{probs} (if \code{drop = TRUE}, default) or a \code{matrix} with
\code{length(x)} columns (if \code{drop = FALSE}). In case of a vectorized distribution
object, a matrix with \code{length(x)} columns containing all possible combinations.
}
\description{
Please see the documentation of \code{\link[=SHASH]{SHASH()}} for some properties
of the SHASH distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

## SHASH() by default uses nu = 1, tau = 1 which
## results in the standard normal distribution
set.seed(6020)
X <- SHASH() # Uses mu = 1, sigma = 0, nu = 1, tau = 1)
x <- random(X, 300)
qqnorm(x); qqline(x, col = 2, lwd = 2)
curve(pdf(X, x), xlim = c(-5, 5), main = paste(X, "density"))

## Calculation of central moments is based on numeric integration,
## thus not being identical to the standard normal distribution
c(mean = mean(x), sd = sd(x))

## Skewed SHASH distribution
X <- SHASH(mu = 7, sigma = 2, nu = c(0.7, 1, 0.7), tau = c(1, 0.7, 0.7))
as.matrix(X)

## Visualization of density functions using different parameters for nu/tau
curve(pdf(X[1], x), xlim = c(0, 20), ylim = c(0, 0.2), main = "Density function")
curve(pdf(X[2], x), xlim = c(0, 20), col = 2, add = TRUE)
curve(pdf(X[3], x), xlim = c(0, 20), col = 4, add = TRUE)

## Visualization of distribution function using different parameters for nu/tau
curve(cdf(X[1], x), xlim = c(0, 20), ylim = 0:1, main = "Distribution function")
curve(cdf(X[2], x), xlim = c(0, 20), col = 2, add = TRUE)
curve(cdf(X[3], x), xlim = c(0, 20), col = 4, add = TRUE)

## Central moments
mean(X)
variance(X)
skewness(X)
kurtosis(X)

## Drawing random values
random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

# note that the cdf() and quantile() functions are inverses
X <- SHASH(mu = 3, sigma = 2, nu = 0.9, tau = 1.2)
cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 7))

}
\seealso{
Other SHASH distribution:
\code{\link[=cdf.SHASH]{cdf.SHASH()}},
\code{\link[=quantile.SHASH]{quantile.SHASH()}}
}
\concept{SHASH distribution}
