% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t.R
\name{pdf.students_t}
\alias{pdf.students_t}
\title{Evaluate the probability mass function of a students_t distribution}
\usage{
\method{pdf}{students_t}(d, x, ...)
}
\arguments{
\item{d}{A \code{students_t} object created by a call to \code{\link[=students_t]{students_t()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Please see the documentation of \code{\link[=students_t]{students_t()}} for some properties
of the students_t distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

# TODO

}
\seealso{
Other students_t distribution: \code{\link{cdf.students_t}},
  \code{\link{quantile.students_t}},
  \code{\link{random.students_t}}
}
\concept{students_t distribution}
