% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentsT.R
\name{pdf.StudentsT}
\alias{pdf.StudentsT}
\alias{log_pdf.StudentsT}
\title{Evaluate the probability mass function of a StudentsT distribution}
\usage{
\method{pdf}{StudentsT}(d, x, ...)

\method{log_pdf}{StudentsT}(d, x, ...)
}
\arguments{
\item{d}{A \code{StudentsT} object created by a call to \code{\link[=StudentsT]{StudentsT()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Please see the documentation of \code{\link[=StudentsT]{StudentsT()}} for some properties
of the StudentsT distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

# TODO

}
\seealso{
Other StudentsT distribution: \code{\link{cdf.StudentsT}},
  \code{\link{quantile.StudentsT}},
  \code{\link{random.StudentsT}}
}
\concept{StudentsT distribution}
