% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{quantile.bernoulli}
\alias{quantile.bernoulli}
\title{Determine quantiles of a bernoulli distribution}
\usage{
\method{quantile}{bernoulli}(d, p, ...)
}
\arguments{
\item{d}{A `bernoulli` object created by a call to [bernoulli()].}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of `p`.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

b <- bernoulli(0.7)
b

random(b, 10)
pdf(b, 1)
cdf(b, 0)
quantile(b, 0.7)

# TODO: looks like I don't quite have quantiles right since the
# inverses don't hold like I'd want

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
