% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{quantile.binomial}
\alias{quantile.binomial}
\title{Determine quantiles of a binomial distribution}
\usage{
\method{quantile}{binomial}(d, p)
}
\arguments{
\item{d}{A `binomial` object created by a call to [binomial()].}

\item{p}{A vector of probabilites.}
}
\value{
A vector of quantiles, one for each element of `p`.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

b <- binomial(10, 0.2)
b

random(b, 10)
pdf(b, 2L)
cdf(b, 4L)
quantile(b, 7)

cdf(b, quantile(b, 7))
quantile(b, cdf(b, 7))

}
