% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{quantile.cauchy}
\alias{quantile.cauchy}
\title{Determine quantiles of a cauchy distribution}
\usage{
\method{quantile}{cauchy}(d, p, ...)
}
\arguments{
\item{d}{A \code{cauchy} object created by a call to \code{\link[=cauchy]{cauchy()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

c <- cauchy(10, 0.2)
c

random(c, 10)
pdf(c, 2)
cdf(c, 2)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
