% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{quantile.chi_square}
\alias{quantile.chi_square}
\title{Determine quantiles of a chi square distribution}
\usage{
\method{quantile}{chi_square}(d, p, ...)
}
\arguments{
\item{d}{A \code{chi_square} object created by a call to \code{\link[=chi_square]{chi_square()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

c <- chi_square(5)
c

random(c, 10)
pdf(c, 2)
cdf(c, 4)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
