% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{quantile.Custom}
\alias{quantile.Custom}
\title{Determine quantiles of a Custom discrete distribution}
\usage{
\method{quantile}{Custom}(d, p, ...)
}
\arguments{
\item{d}{A \code{Custom} object created by a call to \code{\link[=Custom]{Custom()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

X <- Custom(ss = c(1,6,2), probs = c(0.4, 0.1, 0.5))
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))

}
