% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile.default}
\alias{quantile.default}
\title{Use stats::quantile if passing non-distribution object}
\usage{
quantile.default(x, ...)
}
\description{
Use stats::quantile if passing non-distribution object
}
