% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{quantile.exponential}
\alias{quantile.exponential}
\title{Determine quantiles of a exponential distribution}
\usage{
\method{quantile}{exponential}(d, p, ...)
}
\arguments{
\item{d}{A \code{exponential} object created by a call to \code{\link[=exponential]{exponential()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

e <- exponential(5)
e

random(e, 10)
pdf(e, 2)
log_pdf(e, 2)
cdf(e, 4)
quantile(e, 0.7)

cdf(e, quantile(e, 0.7))
quantile(e, cdf(e, 7))

}
