% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_f.R
\name{quantile.fisher_f}
\alias{quantile.fisher_f}
\title{Determine quantiles of an F distribution}
\usage{
\method{quantile}{fisher_f}(d, p, ...)
}
\arguments{
\item{d}{A \code{fisher_f} object created by a call to \code{\link[=fisher_f]{fisher_f()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

f <- fisher_f(5, 10, 0.2)
f

random(f, 10)
pdf(f, 2)
cdf(f, 4)
quantile(f, 0.7)

cdf(f, quantile(f, 0.7))
quantile(f, cdf(f, 7))

}
