% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{quantile.gamma}
\alias{quantile.gamma}
\title{Determine quantiles of a gamma distribution}
\usage{
\method{quantile}{gamma}(d, p, ...)
}
\arguments{
\item{d}{A \code{gamma} object created by a call to \code{\link[=gamma]{gamma()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

g <- gamma(5, 2)
g

random(g, 10)
pdf(g, 2)
log_pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

cdf(g, quantile(g, 0.7))
quantile(g, cdf(g, 7))

}
