% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace.R
\name{quantile.Laplace}
\alias{quantile.Laplace}
\title{Determine quantiles of a Laplace distribution}
\usage{
\method{quantile}{Laplace}(d, p, ...)
}
\arguments{
\item{d}{A \code{Laplace} object created by a call to \code{\link[=Laplace]{Laplace()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

set.seed(27)

X <- Laplace(1, 2)
X

random(X, 10)

pdf(X, c(0.3, 1.7))
log_pdf(X, c(0.3, 1.7))

cdf(X, c(0.3, 1, 1.7))
quantile(X, c(0.25, 0.5, 0.75))

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))

x <- random(X, 100)
fit_mle(X, x)
}
