% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal.R
\name{quantile.log_normal}
\alias{quantile.log_normal}
\title{Determine quantiles of a log_normal distribution}
\usage{
\method{quantile}{log_normal}(d, p, ...)
}
\arguments{
\item{d}{A \code{log_normal} object created by a call to \code{\link[=log_normal]{log_normal()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
Determine quantiles of a log_normal distribution
}
\examples{

ln <- log_normal(0.3, 2)
ln

random(ln, 10)
pdf(ln, 2)
cdf(ln, 4)
quantile(ln, 0.7)

}
\seealso{
Other log_normal distribution: \code{\link{cdf.log_normal}},
  \code{\link{pdf.log_normal}},
  \code{\link{random.log_normal}}
}
\concept{log_normal distribution}
