% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{quantile.negative_binomial}
\alias{quantile.negative_binomial}
\title{Determine quantiles of a negative_binomial distribution}
\usage{
\method{quantile}{negative_binomial}(d, p, ...)
}
\arguments{
\item{d}{A \code{negative_binomial} object created by a call to
\code{\link[=negative_binomial]{negative_binomial()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
Determine quantiles of a negative_binomial distribution
}
\examples{

nb <- negative_binomial(10, 0.3)
nb

random(nb, 10)
pdf(nb, 2)
cdf(nb, 4)
quantile(nb, 0.7)

}
\seealso{
Other negative_binomial distribution: \code{\link{cdf.negative_binomial}},
  \code{\link{pdf.negative_binomial}},
  \code{\link{random.negative_binomial}}
}
\concept{negative_binomial distribution}
