% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{quantile.Pareto}
\alias{quantile.Pareto}
\title{Determine quantiles of a Pareto distribution}
\usage{
\method{quantile}{Pareto}(d, p, ...)
}
\arguments{
\item{d}{A \code{Pareto} object created by a call to \code{\link[=Pareto]{Pareto()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

set.seed(27)

X <- Pareto(1, 2)
X

random(X, 10)

pdf(X, 1.7)
log_pdf(X, 1.7)

cdf(X, 1.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 1.7))

x <- random(X, 100)
fit_mle(X, x)
}
