% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{quantile.poisson}
\alias{quantile.poisson}
\title{Determine quantiles of a poisson distribution}
\usage{
\method{quantile}{poisson}(d, p, ...)
}
\arguments{
\item{d}{A \code{poisson} object created by a call to \code{\link[=poisson]{poisson()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

p <- poisson(2)
p

random(p, 10)
pdf(p, 2)
log_pdf(p, 2)
cdf(p, 4)
quantile(p, 0.7)

cdf(p, quantile(p, 0.7))
quantile(p, cdf(p, 7))

}
