% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{quantile}
\alias{quantile}
\alias{quantile.default}
\title{Find the quantile of a probability distribution}
\usage{
quantile(x, p, ...)

\method{quantile}{default}(x, p, names = FALSE, ...)
}
\arguments{
\item{x}{A probability distribution object such as those created by
a call to \code{\link[=Bernoulli]{Bernoulli()}}, \code{\link[=Beta]{Beta()}}, or \code{\link[=Binomial]{Binomial()}}.}

\item{p}{A vector of probabilities.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
This function allows us to use the \code{\link[stats:quantile]{stats::quantile()}} alongside
custom S3 methods for \code{distributions} objects. When passing arguments
to \code{\link[stats:quantile]{stats::quantile()}}, accepts both
\code{distributions3} style arguments \code{d}, \code{p} and stats style arguments
\code{x}, \code{probs} as well as any combination of the two.
}
\examples{

X <- Normal()

cdf(X, c(0.2, 0.4, 0.6, 0.8))

}
