% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t.R
\name{quantile.students_t}
\alias{quantile.students_t}
\title{Determine quantiles of a students_t distribution}
\usage{
\method{quantile}{students_t}(d, p, ...)
}
\arguments{
\item{d}{A \code{students_t} object created by a call to \code{\link[=students_t]{students_t()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
Please see the documentation of \code{\link[=students_t]{students_t()}} for some properties
of the students_t distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
\code{quantile()}
}
\details{
This function returns the same values that you get from a Z-table. Note
\code{quantile()} is the inverse of \code{cdf()}. Please see the documentation of \code{\link[=students_t]{students_t()}} for some properties
of the students_t distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

# TODO

}
\seealso{
Other students_t distribution: \code{\link{cdf.students_t}},
  \code{\link{pdf.students_t}},
  \code{\link{random.students_t}}
}
\concept{students_t distribution}
