% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentsT.R
\name{quantile.StudentsT}
\alias{quantile.StudentsT}
\title{Determine quantiles of a StudentsT distribution}
\usage{
\method{quantile}{StudentsT}(d, p, ...)
}
\arguments{
\item{d}{A \code{StudentsT} object created by a call to \code{\link[=StudentsT]{StudentsT()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
Please see the documentation of \code{\link[=StudentsT]{StudentsT()}} for some properties
of the StudentsT distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
\code{quantile()}
}
\details{
This function returns the same values that you get from a Z-table. Note
\code{quantile()} is the inverse of \code{cdf()}. Please see the documentation of
\code{\link[=StudentsT]{StudentsT()}} for some properties
of the StudentsT distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

# TODO

}
\seealso{
Other StudentsT distribution: \code{\link{cdf.StudentsT}},
  \code{\link{pdf.StudentsT}},
  \code{\link{random.StudentsT}}
}
\concept{StudentsT distribution}
