% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.R
\name{quantile.uniform}
\alias{quantile.uniform}
\title{Determine quantiles of a continuous uniform  distribution}
\usage{
\method{quantile}{uniform}(d, p, ...)
}
\arguments{
\item{d}{A \code{uniform} object created by a call to \code{\link[=uniform]{uniform()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

b <- uniform(1, 2)
b

random(b, 10)
pdf(b, 0.7)
log_pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
