% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroAdjustedNegbin.R
\name{quantile.ZeroAdjustedNegbin}
\alias{quantile.ZeroAdjustedNegbin}
\title{Determine quantiles of a Zero-Adjusted Negative Binomial distribution}
\usage{
\method{quantile}{ZeroAdjustedNegbin}(x, probs, drop = TRUE, elementwise = NULL, cores = NULL, ...)
}
\arguments{
\item{x}{A \code{ZeroAdjustedNegbin} object created by a call to \code{\link[=ZeroAdjustedNegbin]{ZeroAdjustedNegbin()}}.}

\item{probs}{A vector of probabilities.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{elementwise}{logical. Should each distribution in \code{x} be evaluated
at all elements of \code{probs} (\code{elementwise = FALSE}, yielding a matrix)?
Or, if \code{x} and \code{probs} have the same length, should the evaluation be
done element by element (\code{elementwise = TRUE}, yielding a vector)? The
default of \code{NULL} means that \code{elementwise = TRUE} is used if the
lengths match and otherwise \code{elementwise = FALSE} is used.}

\item{cores}{\code{NULL} or positive integer. If not \code{NULL} we use the C code with \code{cores} threads.}

\item{...}{Arguments to be passed to \code{\link[stats]{qnbinom}}.
Unevaluated arguments will generate a warning to catch mispellings or other
possible errors.}
}
\value{
In case of a single distribution object, either a numeric
vector of length \code{length(probs)} (if \code{drop = TRUE}, default) or a \code{matrix} with
\code{length(probs)} columns (if \code{drop = FALSE}). In case of a vectorized distribution
object, a matrix with \code{length(probs)} columns containing all possible combinations.
}
\description{
Determine quantiles of a Zero-Adjusted Negative Binomial distribution
}
