% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{random.bernoulli}
\alias{random.bernoulli}
\title{Draw a random sample from a bernoulli distribution}
\usage{
\method{random}{bernoulli}(d, n = 1L)
}
\arguments{
\item{d}{A `bernoulli` object created by a call to [bernoulli()].}

\item{n}{The number of samples to draw. Defaults to `1L`.}
}
\value{
An integer vector of zeros and ones of length `n`.
}
\description{
Draw a random sample from a bernoulli distribution
}
\examples{

b <- bernoulli(0.7)
b

random(b, 10)
pdf(b, 1)
cdf(b, 0)
quantile(b, 0.7)

# TODO: looks like I don't quite have quantiles right since the
# inverses don't hold like I'd want

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
