% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{random.beta}
\alias{random.beta}
\title{Draw a random sample from a beta distribution}
\usage{
\method{random}{beta}(d, n = 1L)
}
\arguments{
\item{d}{A `beta` object created by a call to [beta()].}

\item{n}{The number of samples to draw. Defaults to `1L`.}
}
\value{
A numeric vector containing values in `[0, 1]` of length `n`.
}
\description{
Draw a random sample from a beta distribution
}
\examples{

b <- beta(1, 2)
b

random(b, 10)
pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
