% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{random.binomial}
\alias{random.binomial}
\title{Draw a random sample from a binomial distribution}
\usage{
\method{random}{binomial}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{binomial} object created by a call to \code{\link[=binomial]{binomial()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
An integer vector containing values between \code{0} and \code{d$size}
of length \code{n}.
}
\description{
Draw a random sample from a binomial distribution
}
\examples{

b <- binomial(10, 0.2)
b

random(b, 10)
pdf(b, 2L)
log_pdf(b, 2L)
cdf(b, 4L)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 7))

}
