% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{random.cauchy}
\alias{random.cauchy}
\title{Draw a random sample from a cauchy distribution}
\usage{
\method{random}{cauchy}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{cauchy} object created by a call to \code{\link[=cauchy]{cauchy()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from a cauchy distribution
}
\examples{

c <- cauchy(10, 0.2)
c

random(c, 10)
pdf(c, 2)
cdf(c, 2)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
