% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{random.chi_square}
\alias{random.chi_square}
\title{Draw a random sample from a chi square distribution}
\usage{
\method{random}{chi_square}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{chi_square} object created by a call to \code{\link[=chi_square]{chi_square()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from a chi square distribution
}
\examples{

c <- chi_square(5)
c

random(c, 10)
pdf(c, 2)
log_pdf(c, 2)
cdf(c, 4)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
