% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{random.Custom}
\alias{random.Custom}
\title{Draw a random sample from a Custom distribution}
\usage{
\method{random}{Custom}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{Custom} object created by a call to \code{\link[=Custom]{Custom()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector containing values from \code{outcomes} of length \code{n}.
}
\description{
Draw a random sample from a Custom distribution
}
\examples{

X <- Custom(ss = c(1,6,2), probs = c(0.4, 0.1, 0.5))
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))

}
