% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{random.distribution}
\alias{random.distribution}
\title{Inverse transform sampling for random number generation}
\usage{
\method{random}{distribution}(x, n = 1L, drop = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{distribution}.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is
taken to be the number required.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{currently unused.}
}
\value{
A numeric vector of random values if \code{length(d)} equals one
or \code{n = 1L} and \code{drop = TRUE} (default) or a matrix where rows correspond
to the distribution(s) \code{d} whilst the columns contain the random values.
}
\description{
Generates random numbers from a distribution object by mapping
uniform random variates through the target quantile function.
}
\examples{
## Drawing random numbers from a Poisson distribution
## using the inverse transform sampling method
random.distribution(Poisson(3), n = 10)

## Drawing random numbers from a series of Normal distributions
## using the inverse transform sampling method
random.distribution(Normal(1:3, 1:3 / 2), n = 10)

}
