% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{random.Empirical}
\alias{random.Empirical}
\title{Draw a random sample from an Empirical distribution}
\usage{
\method{random}{Empirical}(x, n = 1L, drop = TRUE, ...)
}
\arguments{
\item{x}{A n \code{Empirical} object created by a call to \code{\link[=Empirical]{Empirical()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{currently unused.}
}
\value{
In case of a single distribution object or \code{n = 1}, either a numeric
vector of length \code{n} (if \code{drop = TRUE}, default) or a \code{matrix} with \code{n} columns
(if \code{drop = FALSE}).
}
\description{
Draws \code{n} random values from the empirical distribution with
replacement. Please see the documentation of \code{\link[=Empirical]{Empirical()}} for some
properties of the empircal ensemble distribution.
}
\examples{

set.seed(28)

X <- Empirical(rnorm(50))
X

mean(X)
variance(X)
skewness(X)
kurtosis(X)

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

### example: allowed types/classes of input arguments

## Single vector (will be coerced to numeric)
Y1  <- rnorm(3, mean = -10)
d1 <- Empirical(Y1)
d1
mean(d1)

## Unnamed list of vectors
Y2 <- list(as.character(rnorm(3, mean = -10)),
           runif(6),
           rpois(4, lambda = 15))
d2 <- Empirical(Y2)
d2
mean(d2)

## Named list of vectors
Y3 <- list("Normal"  = as.character(rnorm(3, mean = -10)),
           "Uniform" = runif(6),
           "Poisson" = rpois(4, lambda = 15))
d3 <- Empirical(Y3)
d3
mean(d3)

## Matrix
Y4 <- matrix(rnorm(20), ncol = 5,
             dimnames = list(sprintf("D_\%d", 1:4), sprintf("obs_\%d", 1:5)))
d4 <- Empirical(Y4)
d4

## Data frame
d5 <- Empirical(as.data.frame(t(Y4)))
d5

mean(d5)
variance(d5)
skewness(d5)
kurtosis(d5)

pdf(d5, c(-0.5, 0, 0.5, 1)) # Defaults to elementwise = TRUE
pdf(d5, c(-0.5, 0, 0.5, 1), elementwise = FALSE)

cdf(d5, c(-0.5, 0, 0.5, 1)) # Defaults to elementwise = TRUE
cdf(d5, c(-0.5, 0, 0.5, 1), elementwise = FALSE)

quantile(d5, c(0.2, 0.4, 0.6, 0.8)) # Defaults to elementwise = TRUE
quantile(d5, c(0.2, 0.4, 0.6, 0.8), elementwise = FALSE)

}
\seealso{
Other Empirical distribution:
\code{\link[=Empirical]{Empirical()}},
\code{\link[=cdf.Empirical]{cdf.Empirical()}},
\code{\link[=dempirical]{dempirical()}},
\code{\link[=pdf.Empirical]{pdf.Empirical()}},
\code{\link[=quantile.Empirical]{quantile.Empirical()}},
\code{\link[=support.Empirical]{support.Empirical()}}
}
\concept{Empirical distribution}
