% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{random.exponential}
\alias{random.exponential}
\title{Draw a random sample from a exponential distribution}
\usage{
\method{random}{exponential}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{exponential} object created by a call to \code{\link[=exponential]{exponential()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from a exponential distribution
}
\examples{

e <- exponential(5)
e

random(e, 10)
pdf(e, 2)
log_pdf(e, 2)
cdf(e, 4)
quantile(e, 0.7)

cdf(e, quantile(e, 0.7))
quantile(e, cdf(e, 7))

}
