% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_f.R
\name{random.fisher_f}
\alias{random.fisher_f}
\title{Draw a random sample from an F distribution}
\usage{
\method{random}{fisher_f}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{fisher_f} object created by a call to \code{\link[=fisher_f]{fisher_f()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from an F distribution
}
\examples{

f <- fisher_f(5, 10, 0.2)
f

random(f, 10)
pdf(f, 2)
cdf(f, 4)
quantile(f, 0.7)

cdf(f, quantile(f, 0.7))
quantile(f, cdf(f, 7))

}
