% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{random.gamma}
\alias{random.gamma}
\title{Draw a random sample from a gamma distribution}
\usage{
\method{random}{gamma}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{gamma} object created by a call to \code{\link[=gamma]{gamma()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from a gamma distribution
}
\examples{

g <- gamma(5, 2)
g

random(g, 10)
pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

cdf(g, quantile(g, 0.7))
quantile(g, cdf(g, 7))

}
