% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normal.R
\name{random.log_normal}
\alias{random.log_normal}
\title{Draw a random sample from a log_normal distribution}
\usage{
\method{random}{log_normal}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{log_normal} object created by a call to \code{\link[=log_normal]{log_normal()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
An integer vector of length \code{n}.
}
\description{
Draw a random sample from a log_normal distribution
}
\examples{

ln <- log_normal(0.3, 2)
ln

random(ln, 10)
pdf(ln, 2)
log_pdf(ln, 2)
cdf(ln, 4)
quantile(ln, 0.7)

}
\seealso{
Other log_normal distribution: \code{\link{cdf.log_normal}},
  \code{\link{pdf.log_normal}},
  \code{\link{quantile.log_normal}}
}
\concept{log_normal distribution}
