% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_binomial.R
\name{random.negative_binomial}
\alias{random.negative_binomial}
\title{Draw a random sample from a negative binomial distribution}
\usage{
\method{random}{negative_binomial}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{negative_binomial} object created by a call to
\code{\link[=negative_binomial]{negative_binomial()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
An integer vector of length \code{n}.
}
\description{
Draw a random sample from a negative binomial distribution
}
\examples{

nb <- negative_binomial(10, 0.3)
nb

random(nb, 10)
pdf(nb, 2)
log_pdf(nb, 2)
cdf(nb, 4)
quantile(nb, 0.7)

}
\seealso{
Other negative_binomial distribution: \code{\link{cdf.negative_binomial}},
  \code{\link{pdf.negative_binomial}},
  \code{\link{quantile.negative_binomial}}
}
\concept{negative_binomial distribution}
