% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{random.normal}
\alias{random.normal}
\title{Draw a random sample from a normal distribution}
\usage{
\method{random}{normal}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{normal} object created by a call to \code{\link[=normal]{normal()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from a normal distribution
}
\examples{

n <- normal(5, 2)
n

random(n, 10)
pdf(n, 2)
cdf(n, 4)
quantile(n, 0.7)

cdf(n, quantile(n, 0.7))
quantile(n, cdf(n, 7))

}
