% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{random.Pareto}
\alias{random.Pareto}
\title{Draw a random sample from a Pareto distribution}
\usage{
\method{random}{Pareto}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{Pareto} object created by a call to \code{\link[=Pareto]{Pareto()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from a Pareto distribution
}
\examples{

set.seed(27)

X <- Pareto(1, 2)
X

random(X, 10)

pdf(X, 1.7)
log_pdf(X, 1.7)

cdf(X, 1.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 1.7))

x <- random(X, 100)
fit_mle(X, x)
}
