% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{random.poisson}
\alias{random.poisson}
\title{Draw a random sample from a poisson distribution}
\usage{
\method{random}{poisson}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{poisson} object created by a call to \code{\link[=poisson]{poisson()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Draw a random sample from a poisson distribution
}
\examples{

p <- poisson(2)
p

random(p, 10)
pdf(p, 2)
cdf(p, 4)
quantile(p, 0.7)

cdf(p, quantile(p, 0.7))
quantile(p, cdf(p, 7))

}
