% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{random.SHASH}
\alias{random.SHASH}
\title{Draw a random sample from a SHASH distribution}
\usage{
\method{random}{SHASH}(x, n = 1L, drop = TRUE, ...)
}
\arguments{
\item{x}{A \code{SHASH} object created by a call to \code{\link[=SHASH]{SHASH()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
In case of a single distribution object or \code{n = 1}, either a numeric
vector of length \code{n} (if \code{drop = TRUE}, default) or a \code{matrix} with \code{n} columns
(if \code{drop = FALSE}).
}
\description{
Please see the documentation of \code{\link[=SHASH]{SHASH()}} for some properties
of the SHASH distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

## SHASH(), by default, uses nu = 1, tau = 1 which
## results in the standard normal distribution
set.seed(6020)
X <- SHASH() # Uses mu = 1, sigma = 0, nu = 1, tau = 1)
x <- random(X, 300)
qqnorm(x); qqline(x, col = 2, lwd = 2)
c(mean = mean(x), sd = sd(x))

## TODO(R): Write examples
## set.seed(27)

## X <- SHASH(5, 2)
## X

## mean(X)
## variance(X)
## skewness(X)
## kurtosis(X)

## random(X, 10)

## pdf(X, 2)
## log_pdf(X, 2)

## cdf(X, 4)
## quantile(X, 0.7)

## ### example: calculating p-values for two-sided Z-test

## # here the null hypothesis is H_0: mu = 3
## # and we assume sigma = 2

## # exactly the same as: Z <- SHASH(0, 1)
## Z <- SHASH()

## # data to test
## x <- c(3, 7, 11, 0, 7, 0, 4, 5, 6, 2)
## nx <- length(x)

## # calculate the z-statistic
## z_stat <- (mean(x) - 3) / (2 / sqrt(nx))
## z_stat

## # calculate the two-sided p-value
## 1 - cdf(Z, abs(z_stat)) + cdf(Z, -abs(z_stat))

## # exactly equivalent to the above
## 2 * cdf(Z, -abs(z_stat))

## # p-value for one-sided test
## # H_0: mu <= 3   vs   H_A: mu > 3
## 1 - cdf(Z, z_stat)

## # p-value for one-sided test
## # H_0: mu >= 3   vs   H_A: mu < 3
## cdf(Z, z_stat)

## ### example: calculating a 88 percent Z CI for a mean

## # same `x` as before, still assume `sigma = 2`

## # lower-bound
## mean(x) - quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## # upper-bound
## mean(x) + quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## # equivalent to
## mean(x) + c(-1, 1) * quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## # also equivalent to
## mean(x) + quantile(Z, 0.12 / 2) * 2 / sqrt(nx)
## mean(x) + quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## ### generating random samples and plugging in ks.test()

## set.seed(27)

## # generate a random sample
## ns <- random(SHASH(3, 7), 26)

## # test if sample is SHASH(3, 7)
## ks.test(ns, pnorm, mean = 3, sd = 7)

## # test if sample is gamma(8, 3) using base R pgamma()
## ks.test(ns, pgamma, shape = 8, rate = 3)

## ### MISC

## # note that the cdf() and quantile() functions are inverses
## cdf(X, quantile(X, 0.7))
## quantile(X, cdf(X, 7))

}
