% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t.R
\name{random.students_t}
\alias{random.students_t}
\title{Draw a random sample from a students_t distribution}
\usage{
\method{random}{students_t}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{students_t} object created by a call to \code{\link[=students_t]{students_t()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Please see the documentation of \code{\link[=students_t]{students_t()}} for some properties
of the T distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

# TODO

}
\seealso{
Other students_t distribution: \code{\link{cdf.students_t}},
  \code{\link{pdf.students_t}},
  \code{\link{quantile.students_t}}
}
\concept{students_t distribution}
