% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentsT.R
\name{random.StudentsT}
\alias{random.StudentsT}
\title{Draw a random sample from a StudentsT distribution}
\usage{
\method{random}{StudentsT}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{StudentsT} object created by a call to \code{\link[=StudentsT]{StudentsT()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Please see the documentation of \code{\link[=StudentsT]{StudentsT()}} for some properties
of the T distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
}
\examples{

# TODO

}
\seealso{
Other StudentsT distribution: \code{\link{cdf.StudentsT}},
  \code{\link{pdf.StudentsT}},
  \code{\link{quantile.StudentsT}}
}
\concept{StudentsT distribution}
