% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.R
\name{random.uniform}
\alias{random.uniform}
\title{Draw a random sample from a continuous uniform distribution}
\usage{
\method{random}{uniform}(d, n = 1L, ...)
}
\arguments{
\item{d}{A \code{uniform} object created by a call to \code{\link[=uniform]{uniform()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector containing values in \code{[a, b]} of length \code{n}.
}
\description{
Draw a random sample from a continuous uniform distribution
}
