% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroAdjustedNegbin.R
\name{random.ZeroAdjustedNegbin}
\alias{random.ZeroAdjustedNegbin}
\title{Draw a random sample from a Zero-Adjusted Negative Binomial distribution}
\usage{
\method{random}{ZeroAdjustedNegbin}(x, n = 1L, drop = TRUE, cores = NULL, ...)
}
\arguments{
\item{x}{A \code{ZeroAdjustedNegbin} object created by a call to \code{\link[=ZeroAdjustedNegbin]{ZeroAdjustedNegbin()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{cores}{\code{NULL} or positive integer. TODO(R): Just a development option.
If not \code{NULL} we use the C code with \code{cores} threads.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
In case of a single distribution object or \code{n = 1}, either a numeric
vector of length \code{n} (if \code{drop = TRUE}, default) or a \code{matrix} with \code{n} columns
(if \code{drop = FALSE}).
}
\description{
Please see the documentation of \code{\link[=ZeroAdjustedNegbin]{ZeroAdjustedNegbin()}} for some properties
of the Zero-Adjusted Negative Binomial distribution.
}
