% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{dshash}
\alias{dshash}
\alias{pshash}
\alias{qshash}
\alias{rshash}
\title{The Sinh-Arcsinh (SHASH) distribution}
\usage{
dshash(x, mu = 0, sigma = 1, nu = 1, tau = 1, log = FALSE)

pshash(q, mu = 0, sigma = 1, nu = 1, tau = 1, lower.tail = TRUE, log.p = FALSE)

qshash(p, mu = 0, sigma = 1, nu = 1, tau = 1, lower.tail = TRUE, log.p = FALSE)

rshash(n, mu = 0, sigma = 1, nu = 1, tau = 1)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{log, log.p}{logical indicating whether probabilities p are given as log(p).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical indicating whether probabilities are \eqn{P[X \le x]} (lower tail) or \eqn{P[X > x]} (upper tail).}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}

\item{lambda}{vector of (non-negative) Poisson parameters.}
}
\description{
Density, distribution function, quantile function, and random
generation for the sinh-arcsinh distribution with four parameters
\code{mu}, \code{sigma}, \code{nu}, and \code{tau}.
}
\details{
The Sinh-Arcsinh generalizes the Normal distribution by separately
controlling location, scale, skewness, and tail-heaviness and can thus produce a
wide range of shapes. Using \code{nu = 1} and \code{tau = 1} results in a
normal distribution.

All functions follow the usual conventions of d/p/q/r functions in base R.
}
\examples{
## theoretical probabilities for a zero-truncated Poisson distribution
x <- 0:8
p <- dztpois(x, lambda = 2.5)
plot(x, p, type = "h", lwd = 2)

## corresponding empirical frequencies from a simulated sample
set.seed(0)
y <- rztpois(500, lambda = 2.5)
hist(y, breaks = -1:max(y) + 0.5)

}
\seealso{
\code{\link{ZTPoisson}}, \code{\link{dpois}}
}
\keyword{distribution}
