% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{StatAUC}
\alias{StatAUC}
\title{Stat for Area Under Curve}
\format{
An object of class \code{StatAUC} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatAUC
}
\description{
Stat for Area Under Curve
}
\keyword{datasets}
