% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{suff_stat.binomial}
\alias{suff_stat.binomial}
\title{Compute the sufficient statistics for the binomial distribution from data}
\usage{
\method{suff_stat}{binomial}(d, x, ...)
}
\arguments{
\item{d}{A \code{binomial} object.}

\item{x}{A vector of zeroes and ones.}
}
\value{
A named list of the sufficient statistics of the binomial distribution
\describe{
\item{\code{successes}}{The total number of successful trials}
\item{\code{experiments}}{The number of experiments run}
\item{\code{trials}}{The number of trials run per experiment}
}
}
\description{
Compute the sufficient statistics for the binomial distribution from data
}
