% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{suff_stat.Erlang}
\alias{suff_stat.Erlang}
\title{Compute the sufficient statistics for an Erlang distribution from data}
\usage{
\method{suff_stat}{Erlang}(d, x, ...)
}
\arguments{
\item{d}{A \code{Erlang} object created by a call to \code{\link[=Erlang]{Erlang()}}.}

\item{x}{A vector to fit the Erlang distribution to.}

\item{...}{Unused.}
}
\value{
a \code{Erlang} object
}
\description{
\itemize{
\item \code{sum}: The sum of the data.
\itemize{
\item \code{log_sum}: The log of the sum of the data.
\item \code{samples}: The number of samples in the data.
}
}
}
