% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric.R
\name{suff_stat.Geometric}
\alias{suff_stat.Geometric}
\title{Compute the sufficient statistics for the Geometric distribution from data}
\usage{
\method{suff_stat}{Geometric}(d, x, ...)
}
\arguments{
\item{d}{A \code{Geometric} object.}

\item{x}{A vector of zeroes and ones.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the Geometric
distribution:
\itemize{
\item \code{trials}: The total number of trials ran until the first success.
\item \code{experiments}: The number of experiments run.
}
}
\description{
Compute the sufficient statistics for the Geometric distribution from data
}
