% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace.R
\name{suff_stat.Laplace}
\alias{suff_stat.Laplace}
\title{Compute the sufficient statistics for a Laplace distribution from data}
\usage{
\method{suff_stat}{Laplace}(d, x, ...)
}
\arguments{
\item{d}{A \code{Laplace} object created by a call to \code{\link[=Laplace]{Laplace()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the Laplace
distribution:
\itemize{
\item \code{theta}: The sample median.
\item \code{sigma}: The sample mean absolute deviation from the sample median.
\item \code{samples}: The number of samples in the data.
}
}
\description{
Compute the sufficient statistics for a Laplace distribution from data
}
