% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{suff_stat.LogNormal}
\alias{suff_stat.LogNormal}
\title{Compute the sufficient statistics for a Log-normal distribution from data}
\usage{
\method{suff_stat}{LogNormal}(d, x, ...)
}
\arguments{
\item{d}{A \code{LogNormal} object created by a call to \code{\link[=LogNormal]{LogNormal()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the normal distribution:
\itemize{
\item \code{mu}: The sample mean of the log of the data.
\item \code{sigma}: The sample standard deviation of the log of the data.
\item \code{samples}: The number of samples in the data.
}
}
\description{
Compute the sufficient statistics for a Log-normal distribution from data
}
