% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto.R
\name{suff_stat.Pareto}
\alias{suff_stat.Pareto}
\title{Compute the sufficient statistics for a Pareto distribution from data}
\usage{
\method{suff_stat}{Pareto}(d, x, ...)
}
\arguments{
\item{d}{A \code{Pareto} object created by a call to \code{\link[=Pareto]{Pareto()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the Pareto
distribution:
\itemize{
\item \code{minimum}: The sample minimum.
\item \code{mean_log}: The sample mean of the natural logs of the data.
\item \code{samples}: The number of samples in the data.
}
}
\description{
Compute the sufficient statistics for a Pareto distribution from data
}
