% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernoulli.R
\name{support.Bernoulli}
\alias{support.Bernoulli}
\title{Return the support of the Bernoulli distribution}
\usage{
\method{support}{Bernoulli}(d)
}
\arguments{
\item{d}{An \code{Bernoulli} object created by a call to \code{\link[=Bernoulli]{Bernoulli()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Bernoulli distribution
}
