% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta.R
\name{support.Beta}
\alias{support.Beta}
\title{Return the support of the Beta distribution}
\usage{
\method{support}{Beta}(d)
}
\arguments{
\item{d}{An \code{Beta} object created by a call to \code{\link[=Beta]{Beta()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Beta distribution
}
