% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{support.Exponential}
\alias{support.Exponential}
\title{Return the support of the Exponential distribution}
\usage{
\method{support}{Exponential}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{Exponential} object created by a call to \code{\link[=Exponential]{Exponential()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Exponential distribution
}
