% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherF.R
\name{support.FisherF}
\alias{support.FisherF}
\title{Return the support of the FisherF distribution}
\usage{
\method{support}{FisherF}(d)
}
\arguments{
\item{d}{An \code{FisherF} object created by a call to \code{\link[=FisherF]{FisherF()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the FisherF distribution
}
