% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{support.Gamma}
\alias{support.Gamma}
\title{Return the support of the Gamma distribution}
\usage{
\method{support}{Gamma}(d)
}
\arguments{
\item{d}{An \code{Gamma} object created by a call to \code{\link[=Gamma]{Gamma()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Gamma distribution
}
