% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric.R
\name{support.Geometric}
\alias{support.Geometric}
\title{Return the support of the Geometric distribution}
\usage{
\method{support}{Geometric}(d)
}
\arguments{
\item{d}{An \code{Geometric} object created by a call to \code{\link[=Geometric]{Geometric()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Geometric distribution
}
