% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdleNegativeBinomial.R
\name{support.HurdleNegativeBinomial}
\alias{support.HurdleNegativeBinomial}
\title{Return the support of the hurdle negative binomial distribution}
\usage{
\method{support}{HurdleNegativeBinomial}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{HurdleNegativeBinomial} object created by a call to \code{\link[=HurdleNegativeBinomial]{HurdleNegativeBinomial()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the hurdle negative binomial distribution
}
