% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurdlePoisson.R
\name{support.HurdlePoisson}
\alias{support.HurdlePoisson}
\title{Return the support of the hurdle Poisson distribution}
\usage{
\method{support}{HurdlePoisson}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{HurdlePoisson} object created by a call to \code{\link[=HurdlePoisson]{HurdlePoisson()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the hurdle Poisson distribution
}
