% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\name{support.Logistic}
\alias{support.Logistic}
\title{Return the support of the Logistic distribution}
\usage{
\method{support}{Logistic}(d)
}
\arguments{
\item{d}{An \code{Logistic} object created by a call to \code{\link[=Logistic]{Logistic()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Logistic distribution
}
