% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{support.LogNormal}
\alias{support.LogNormal}
\title{Return the support of the LogNormal distribution}
\usage{
\method{support}{LogNormal}(d)
}
\arguments{
\item{d}{An \code{LogNormal} object created by a call to \code{\link[=LogNormal]{LogNormal()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the LogNormal distribution
}
