% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegativeBinomial.R
\name{support.NegativeBinomial}
\alias{support.NegativeBinomial}
\title{Return the support of the NegativeBinomial distribution}
\usage{
\method{support}{NegativeBinomial}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{NegativeBinomial} object created by a call to \code{\link[=NegativeBinomial]{NegativeBinomial()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the NegativeBinomial distribution
}
