% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\name{support.Poisson}
\alias{support.Poisson}
\title{Return the support of the Poisson distribution}
\usage{
\method{support}{Poisson}(d)
}
\arguments{
\item{d}{An \code{Poisson} object created by a call to \code{\link[=Poisson]{Poisson()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Poisson distribution
}
