% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonBinomial.R
\name{support.PoissonBinomial}
\alias{support.PoissonBinomial}
\title{Return the support of the PoissonBinomial distribution}
\usage{
\method{support}{PoissonBinomial}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{A \code{PoissonBinomial} object created by a call to \code{\link[=PoissonBinomial]{PoissonBinomial()}}.}

\item{drop}{logical. Shoul the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the PoissonBinomial distribution
}
