% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentsT.R
\name{support.StudentsT}
\alias{support.StudentsT}
\title{Return the support of the StudentsT distribution}
\usage{
\method{support}{StudentsT}(d)
}
\arguments{
\item{d}{An \code{StudentsT} object created by a call to \code{\link[=StudentsT]{StudentsT()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the StudentsT distribution
}
