% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weibull.R
\name{support.Weibull}
\alias{support.Weibull}
\title{Return the support of the Weibull distribution}
\usage{
\method{support}{Weibull}(d)
}
\arguments{
\item{d}{An \code{Weibull} object created by a call to \code{\link[=Weibull]{Weibull()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Weibull distribution
}
