% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZTNegativeBinomial.R
\name{support.ZTNegativeBinomial}
\alias{support.ZTNegativeBinomial}
\title{Return the support of the zero-truncated negative binomial distribution}
\usage{
\method{support}{ZTNegativeBinomial}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{ZTNegativeBinomial} object created by a call to \code{\link[=ZTNegativeBinomial]{ZTNegativeBinomial()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the zero-truncated negative binomial distribution
}
