% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZTPoisson.R
\name{support.ZTPoisson}
\alias{support.ZTPoisson}
\title{Return the support of the zero-truncated Poisson distribution}
\usage{
\method{support}{ZTPoisson}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{ZTPoisson} object created by a call to \code{\link[=ZTPoisson]{ZTPoisson()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the zero-truncated Poisson distribution
}
