% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{variance}
\alias{variance}
\alias{skewness}
\alias{kurtosis}
\title{Compute the moments of a probability distribution}
\usage{
variance(d, ...)

skewness(d, ...)

kurtosis(d, ...)
}
\arguments{
\item{d}{A probability distribution object such as those created by
a call to \code{\link[=Bernoulli]{Bernoulli()}}, \code{\link[=Beta]{Beta()}}, or \code{\link[=Binomial]{Binomial()}}.}

\item{...}{Further arguments passed to or from other methods (currently not used).}
}
\value{
A numeric scalar
}
\description{
Compute the moments of a probability distribution
}
