% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_dpqr}
\alias{apply_dpqr}
\title{Helper Function for Vectorization of Distribution Objects}
\usage{
apply_dpqr(d, FUN, at, drop = TRUE, type = NULL, ...)
}
\arguments{
\item{d}{A \code{distributions3} object.}

\item{FUN}{Function to be computed. Function should be of type \code{FUN(at, d)}, see details below.}

\item{at}{Specification of values at which \code{FUN} should be
evaluated, typically a numeric vector but possibly also a matrix or data
frame.}

\item{drop}{logical. Should the result be simplified to an unnamed vector if possible?}

\item{type}{Character string used for naming, typically on of \code{"density"}, \code{"logLik"},
\code{"probability"}, \code{"quantile"}, and \code{"random"}. Beware \code{"random"} is processed internally as a special case.}

\item{...}{Arguments to be passed to  \code{FUN()}.}
}
\description{
Helper Function for Vectorization of Distribution Objects
}
