% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bernoulli}
\alias{bernoulli}
\title{Create a bernoulli distribution}
\usage{
bernoulli(p = 0.5)
}
\arguments{
\item{p}{The success probability for the distribution. \code{p} can be any
value in \code{[0, 1]}, and defaults to \code{0.5}.}
}
\value{
A \code{bernoulli} object.
}
\description{
Bernoulli distributions are used to represent events like coin flips
when there is single trial that is either successful or unsuccessful.
The bernoulli distribution is a special case of the \code{\link[=binomial]{binomial()}}
distribution with \code{n = 1}.
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the
math will render nicely.

In the following, let \eqn{X} be a Bernoulli random variable with parameter
\code{p} = \eqn{p}. Some textbooks also define \eqn{q = 1 - p}, or use
\eqn{\pi} instead of \eqn{p}.

\strong{Support}: \eqn{0, 1}

\strong{Mean}: \eqn{p}

\strong{Variance}: \eqn{p \cdot (1 - p) = p \cdot q}

\strong{Probability mass function (p.m.f)}:

\deqn{P(X = x) = p^x (1 - p)^{1-x} = p^x q^{1-x}}

\strong{Cumulative distribution function (c.d.f)}:

TODO ugh formatting

\strong{Moment generating function (m.g.f)}:

\deqn{\mathbb{E}(e^{tX}) = (1 - p) + p e^t}
}
\examples{

b <- bernoulli(0.7)
b

random(b, 10)
pdf(b, 1)
cdf(b, 0)
quantile(b, 0.7)

# TODO: looks like I don't quite have quantiles right since the
# inverses don't hold like I'd want

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
\seealso{
Other discrete distributions: \code{\link{binomial}}
}
\concept{discrete distributions}
