% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bernoulli}
\alias{bernoulli}
\title{Create a bernoulli distribution}
\usage{
bernoulli(p = 0.5)
}
\arguments{
\item{p}{The success probability for the distribution. `p` can be any
value in `[0, 1]`, and defaults to `0.5`.}
}
\value{
A `bernoulli` object.
}
\description{
Create a bernoulli distribution
}
\examples{

b <- bernoulli(0.7)
b

random(b, 10)
pdf(b, 1)
cdf(b, 0)
quantile(b, 0.7)

# TODO: looks like I don't quite have quantiles right since the
# inverses don't hold like I'd want

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
\seealso{
Other discrete distributions: \code{\link{binomial}}
}
\concept{discrete distributions}
