% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta}
\alias{beta}
\title{Create a beta distribution}
\usage{
beta(alpha = 1, beta = 1)
}
\arguments{
\item{alpha}{The alpha parameter. \code{alpha} can be any value strictly
greater than zero. Defaults to \code{1}.}

\item{beta}{The beta parameter. \code{beta} can be any value strictly
greater than zero. Defaults to \code{1}.}
}
\value{
A \code{beta} object.
}
\description{
Create a beta distribution
}
\examples{

b <- beta(1, 2)
b

random(b, 10)
pdf(b, 0.7)
log_pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
\seealso{
Other continuous distributions: \code{\link{cauchy}},
  \code{\link{chi_square}}, \code{\link{exponential}},
  \code{\link{fisher_f}}, \code{\link{gamma}},
  \code{\link{normal}}, \code{\link{poisson}},
  \code{\link{students_t}}
}
\concept{continuous distributions}
