% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binomial}
\alias{binomial}
\title{Create a binomial distribution}
\usage{
binomial(size, p)
}
\arguments{
\item{size}{The number of trials. Must be an integer greater than or equal
to one. When \code{size = 1L}, the binomial distribution reduces to the
bernoulli distribution. Oftened called \code{n} in textbooks.}

\item{p}{The success probability for a given trial. \code{p} can be any
value in \code{[0, 1]}, and defaults to \code{0.5}.
greater than zero. Defaults to \code{1}.}
}
\value{
A \code{binomial} object.
}
\description{
Create a binomial distribution
}
\examples{

b <- binomial(10, 0.2)
b

random(b, 10)
pdf(b, 2L)
cdf(b, 4L)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 7))

}
\seealso{
Other discrete distributions: \code{\link{bernoulli}}
}
\concept{discrete distributions}
