% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{cauchy}
\alias{cauchy}
\title{Create a cauchy distribution}
\usage{
cauchy(location = 0, scale = 1)
}
\arguments{
\item{location}{The location parameter. Can be any real number. Defaults
to \code{0}.}

\item{scale}{The scale parameter. Must be greater than zero (?). Defaults
to \code{1}.}
}
\value{
A \code{cauchy} object.
}
\description{
Note that the cauchy distribution is the student's t distribution
with one degree of freedom. The cauchy distribution does not have
a well defined mean or variance.
}
\examples{

c <- cauchy(10, 0.2)
c

random(c, 10)
pdf(c, 2)
cdf(c, 2)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{chi_square}}, \code{\link{exponential}},
  \code{\link{fisher_f}}, \code{\link{gamma}},
  \code{\link{normal}}, \code{\link{poisson}}
}
\concept{continuous distributions}
