% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{cauchy}
\alias{cauchy}
\title{Create a cauchy distribution}
\usage{
cauchy(location = 0, scale = 1)
}
\arguments{
\item{location}{The location parameter. Can be any real number. Defaults
to \code{0}.}

\item{scale}{The scale parameter. Must be greater than zero (?). Defaults
to \code{1}.}
}
\value{
A \code{cauchy} object.
}
\description{
Note that the cauchy distribution is the student's t distribution
with one degree of freedom. The cauchy distribution does not have
a well defined mean or variance. Cauchy distributions often appear
as priors in Bayesian contexts due to their heavy tails.
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a Cauchy variable with mean
\code{location =} \eqn{x_0} and \code{scale} = \eqn{\gamma}.

\strong{Support}: \eqn{\mathbb{R}}{R}, the set of all real numbers

\strong{Mean}: Undefined.

\strong{Variance}: Undefined.

\strong{Probability density function (p.d.f)}:

\deqn{
    f(x) = \frac{1}{\pi \gamma \left[1 + \left(\frac{x - x_0}{\gamma} \right)^2 \right]}
  }{
    f(x) = 1 / (\pi \gamma (1 + ((x - x_0) / \gamma)^2)
  }

\strong{Cumulative distribution function (c.d.f)}:

\deqn{
    F(t) = \frac{1}{\pi} \arctan \left( \frac{t - x_0}{\gamma} \right) +
      \frac{1}{2}
  }{
    F(t) = arctan((t - x_0) / \gamma) / \pi + 1/2
  }

\strong{Moment generating function (m.g.f)}:

Does not exist.
}
\examples{

c <- cauchy(10, 0.2)
c

random(c, 10)
pdf(c, 2)
log_pdf(c, 2)
cdf(c, 2)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{chi_square}}, \code{\link{exponential}},
  \code{\link{fisher_f}}, \code{\link{gamma}},
  \code{\link{normal}}, \code{\link{poisson}},
  \code{\link{students_t}}, \code{\link{uniform}}
}
\concept{continuous distributions}
