% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{chi_square}
\alias{chi_square}
\title{Create a chi square distribution}
\usage{
chi_square(df)
}
\arguments{
\item{df}{Degrees of freedom. Must be positive.}
}
\value{
A \code{chi_square} object.
}
\description{
Chi square distributions show up often in frequentist settings
as the sampling distribution of test statistics, especially
in maximum likelihood estimation settings. The chi square
distribution with a single degree of freedom is equivalent
to a squared standard normal distribution. Summing chi square
distributions results in a new chi square distribution with
added degrees of freedom. A chi square distribution is a special
case of the gamma distribution with \code{shape = 2} (double check this).
TODO: add relationship to F and T distributions.
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a \eqn{\chi^2} random variable with
\code{df} = \eqn{k}.

\strong{Support}: \eqn{\mathbb{R}^+}{R^+}, the set of positive real numbers

\strong{Mean}: \eqn{k}

\strong{Variance}: \eqn{2k}

\strong{Probability density function (p.d.f)}:

\deqn{
    f(x) = \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2}
  }{
    f(x) = 1 / (2 \pi \sigma^2) exp(-(x - \mu)^2 / (2 \sigma^2))
  }

\strong{Cumulative distribution function (c.d.f)}:

The cumulative distribution function has the form

\deqn{
    F(t) = \int_{-\infty}^t \frac{1}{\sqrt{2 \pi \sigma^2}} e^{-(x - \mu)^2 / 2 \sigma^2} dx
  }{
    F(t) = integral_{-\infty}^t 1 / (2 \pi \sigma^2) exp(-(x - \mu)^2 / (2 \sigma^2)) dx
  }

but this integral does not have a closed form solution and must be
approximated numerically. The c.d.f. of a standard normal is sometimes
called the "error function". The notation \eqn{\Phi(t)} also stands
for the c.d.f. of a standard normal evaluated at \eqn{t}. Z-tables
list the value of \eqn{\Phi(t)} for various \eqn{t}.

\strong{Moment generating function (m.g.f)}:

\deqn{
    \mathbb{E}(e^{tX}) = e^{\mu t + \sigma^2 t^2 / 2}
  }{
    E(e^(tX)) = e^(\mu t + \sigma^2 t^2 / 2)
  }
}
\examples{

c <- chi_square(5)
c

random(c, 10)
pdf(c, 2)
log_pdf(c, 2)
cdf(c, 4)
quantile(c, 0.7)

cdf(c, quantile(c, 0.7))
quantile(c, cdf(c, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{exponential}},
  \code{\link{fisher_f}}, \code{\link{gamma}},
  \code{\link{normal}}, \code{\link{poisson}},
  \code{\link{students_t}}, \code{\link{uniform}}
}
\concept{continuous distributions}
