% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{ChiSquare}
\alias{ChiSquare}
\title{Create a chi-square distribution}
\usage{
ChiSquare(df)
}
\value{
A \code{ChiSquare} object.
}
\description{
Chi-square distributions show up often in frequentist settings
as the sampling distribution of test statistics, especially
in maximum likelihood estimation settings. The chi-square
distribution with a single degree of freedom is equivalent
to a squared standard normal distribution, and also to a Gamma distribution
with \eqn{shape = 1/2} and \eqn{rate = 1/2} chi squar\eqn{n}
e
results in a new chi-square distribution with added degrees of
freedom, given by \eqn{\sum_{i = 1}^{n} n_i}. Thereby, summing up
normal random variables (which are \eqn{ChiSq(1)}) will result in a
\eqn{ChiSq(n)}. A chi-square distribution with \eqn{n} degrees of freedom
is also a special case of the Gamma distribution with
\eqn{shape = n/2} and \eqn{rate = 1/2} (see \code{\link[=Gamma]{Gamma()}}).
}
\examples{

X <- ChiSquare(5)
X

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 7))

}
\seealso{
Other continuous distributions: \code{\link{Beta}},
  \code{\link{Cauchy}}, \code{\link{Exponential}},
  \code{\link{FisherF}}, \code{\link{Gamma}},
  \code{\link{LogNormal}}, \code{\link{Logistic}},
  \code{\link{Normal}}, \code{\link{StudentsT}},
  \code{\link{Tukey}}, \code{\link{Uniform}},
  \code{\link{Weibull}}
}
\concept{continuous distributions}
