% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{dempirical}
\alias{dempirical}
\alias{pempirical}
\alias{qempirical}
\alias{rempirical}
\title{The Empirical Distribution}
\usage{
dempirical(x, y, log = FALSE, na.rm = FALSE, method = NULL, ...)

pempirical(q, y, lower.tail = TRUE, log.p = FALSE, na.rm = TRUE)

qempirical(p, y, lower.tail = TRUE, log.p = FALSE, na.rm = TRUE, ...)

rempirical(n, y, na.rm = TRUE)
}
\arguments{
\item{x}{vector of finite quantiles.}

\item{y}{vector of observations of the empirical distribution with two or
more non-missing finite values.}

\item{log, log.p}{logical indicating whether probabilities p are given as log(p)
(both default to \code{FALSE}).}

\item{na.rm}{logical indicating whether missing values (\code{NA}) are stripped
before computation, defaults to \code{TRUE}.}

\item{method}{\code{NULL} or one of \code{"hist"} or \code{"density"}. If \code{NULL}, \code{y} is
considered a random variable from a discrete empirical distribution.
Method \code{"hist"} and \code{"density"} approximate a 'continuous' distribution
based on the empirical sample \code{y}.}

\item{...}{allows to forward arguments to \code{\link[graphics]{hist}}, \code{\link[=density]{density()}}
and \code{\link[base]{apply}}/\code{\link[base]{sapply}} when calling
\code{\link[=dempirical]{dempirical()}} or sample \code{\link[stats]{quantile}} function when calling
\code{\link[=qempirical]{qempirical()}}. Else currently unused.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical indicating whether probabilities are \eqn{P[X \le x]}
(lower tail) or \eqn{P[X > x]} (upper tail).}

\item{p}{numeric vector of probabilities (\verb{[0, 1]}).}
}
\description{
Density (point mass), distribution, quantile function, as well as
a random generation for the Empirical distribution.
}
\details{
All functions follow the usual conventions of d/p/q/r functions in base R. In
particular, all four functions for the Empirical distribution call
the corresponding \verb{*empirical} functions.
}
\seealso{
Other Empirical distribution:
\code{\link[=Empirical]{Empirical()}},
\code{\link[=cdf.Empirical]{cdf.Empirical()}},
\code{\link[=pdf.Empirical]{pdf.Empirical()}},
\code{\link[=quantile.Empirical]{quantile.Empirical()}},
\code{\link[=random.Empirical]{random.Empirical()}},
\code{\link[=support.Empirical]{support.Empirical()}}
}
\concept{Empirical distribution}
