% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-package.R
\docType{package}
\name{distributions-package}
\alias{distributions}
\alias{distributions-package}
\title{distributions: Probability Distributions as S3 Objects}
\description{
Tools to create and manipulate probability
    distributions using S3.  Generics random(), pdf(), cdf() and
    quantile() provide replacements for base R's r/d/p/q style functions.
    Functions and arguments have been named carefully to minimize
    confusion for students in intro stats courses. The documentation for
    each distribution contains detailed mathematical notes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/distributions}
  \item Report bugs at \url{https://github.com/alexpghayes/distributions/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (0000-0002-4985-5160)

}
\keyword{internal}
