% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions3-package.R
\docType{package}
\name{distributions3-package}
\alias{distributions3}
\alias{distributions3-package}
\title{distributions3: Probability Distributions as S3 Objects}
\description{
Tools to create and manipulate probability distributions using S3. Generics pdf(), cdf(), quantile(), and random() provide replacements for base R's d/p/q/r style functions. Functions and arguments have been named carefully to minimize confusion for students in intro stats courses. The documentation for each distribution contains detailed mathematical notes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zeileis/distributions3}
  \item \url{https://zeileis.github.io/distributions3/}
  \item Report bugs at \url{https://github.com/zeileis/distributions3/issues}
}

}
\author{
\strong{Maintainer}: Achim Zeileis \email{Achim.Zeileis@R-project.org} (\href{https://orcid.org/0000-0003-0918-3766}{ORCID})

Authors:
\itemize{
  \item Achim Zeileis \email{Achim.Zeileis@R-project.org} (\href{https://orcid.org/0000-0003-0918-3766}{ORCID})
  \item Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})
  \item Ralph Moller-Trane (\href{https://orcid.org/0000-0001-7824-1785}{ORCID})
  \item Daniel Jordan \email{dandermotj@gmail.com}
  \item Paul Northrop \email{p.northrop@ucl.ac.uk} (\href{https://orcid.org/0000-0002-1992-4882}{ORCID})
  \item Moritz N. Lang \email{moritz.n.lang@gmail.com} (\href{https://orcid.org/0000-0002-2533-9903}{ORCID})
  \item Reto Stauffer \email{reto.stauffer@uibk.ac.at} (\href{https://orcid.org/0000-0002-3798-5507}{ORCID})
}

Other contributors:
\itemize{
  \item Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID}) [contributor]
  \item Bruna Wundervald \email{brunadaviesw@gmail.com} (\href{https://orcid.org/0000-0001-8163-220X}{ORCID}) [contributor]
  \item Alessandro Gasparini \email{alessandro.gasparini@ki.se} (\href{https://orcid.org/0000-0002-8319-7624}{ORCID}) [contributor]
}

}
\keyword{internal}
