% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{Empirical}
\alias{Empirical}
\alias{mean.Empirical}
\alias{variance.Empirical}
\alias{skewness.Empirical}
\alias{kurtosis.Empirical}
\alias{random.Empirical}
\alias{pdf.Empirical}
\alias{log_pdf.Empirical}
\alias{quantile.Empirical}
\alias{support.Empirical}
\title{Create an Empirical distribution}
\usage{
Empirical(x)

\method{mean}{Empirical}(x, ...)

\method{variance}{Empirical}(x, ...)

\method{skewness}{Empirical}(x, type = 1L, ...)

\method{kurtosis}{Empirical}(x, type = 3L, ...)

\method{random}{Empirical}(x, n = 1L, drop = TRUE, ...)

\method{pdf}{Empirical}(d, x, drop = TRUE, elementwise = NULL, ...)

\method{log_pdf}{Empirical}(d, x, drop = TRUE, elementwise = NULL, ...)

\method{quantile}{Empirical}(x, probs, drop = TRUE, elementwise = NULL, ...)

\method{support}{Empirical}(d, drop = TRUE, ...)
}
\arguments{
\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{currently not used.}

\item{type}{integer between \code{1L} and \code{3L} (default) selecting one of three
algorithms. See Details for more information.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{d}{An \code{Empirical} object created by a call to \code{\link[=Empirical]{Empirical()}}.}

\item{elementwise}{logical. Should each distribution in \code{x} be evaluated
at all elements of \code{probs} (\code{elementwise = FALSE}, yielding a matrix)?
Or, if \code{x} and \code{probs} have the same length, should the evaluation be
done element by element (\code{elementwise = TRUE}, yielding a vector)? The
default of \code{NULL} means that \code{elementwise = TRUE} is used if the
lengths match and otherwise \code{elementwise = FALSE} is used.}

\item{probs}{A vector of probabilities.}
}
\value{
An \code{Empirical} object.

In case of a single distribution object or \code{n = 1}, either a numeric
vector of length \code{n} (if \code{drop = TRUE}, default) or a \code{matrix} with \code{n} columns
(if \code{drop = FALSE}).

In case of a single distribution object, either a numeric
vector of length \code{probs} (if \code{drop = TRUE}, default) or a \code{matrix} with
\code{length(x)} columns (if \code{drop = FALSE}). In case of a vectorized distribution
object, a matrix with \code{length(x)} columns containing all possible combinations.

In case of a single distribution object, either a numeric
vector of length \code{probs} (if \code{drop = TRUE}, default) or a \code{matrix} with
\code{length(probs)} columns (if \code{drop = FALSE}). In case of a vectorized
distribution object, a matrix with \code{length(probs)} columns containing all
possible combinations.

In case of a single distribution object, a numeric vector of length 2
with the minimum and maximum value of the support (if \code{drop = TRUE}, default)
or a \code{matrix} with 2 columns. In case of a vectorized distribution object, a
matrix with 2 columns containing all minima and maxima.
}
\description{
An empirical distribution consists of a series of \code{N} observations
out of a typically unknown distribution, i.e., a random sample \eqn{X}.

Draws \code{n} random values from the empirical ensemble
with replacement.

Please see the documentation of \code{\link[=Empirical]{Empirical()}} for some properties
of the empircal ensemble distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.

Please see the documentation of \code{\link[=Empirical]{Empirical()}} for some properties
of the Empirical distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
\code{quantile()}

TODO(RETO): Check description
}
\details{
The creation function \code{\link[=Empirical]{Empirical()}} allows for a variety of different objects
as main input \code{x}.
\itemize{
\item Vector: Assumes that the vector contains a series of observations from one
empirical distribution.
\item List (named or unnamed) of vectors: Each element in the list describes
one empirical distribution defined by the numeric values in each of the vectors.
\item Data frame: Each column (variable) in the data frame describes one empirical
distribution.
\item Matrix: Each row corresponds to one empirical distribution, whilst
the columns contain the individual observations.
}

Missing values are allowed, however, each distribution requires at least two
finite observations (\code{-Inf}/\code{Inf} is replaced by \code{NA}).
Certain types of moments (see \code{\link[=skewness.Empirical]{skewness.Empirical()}}, \code{\link[=kurtosis.Empirical]{kurtosis.Empirical()}})
require at least three or four finite observations.

\strong{Support}: \eqn{R}, the set of all real numbers
\itemize{
\item Expectation: \deqn{E(x) = \text{mean(x)} = \frac{1}{N} \sum_{i=1}^{N} x_i}{mean(x) = 1 / N * sum(x)}
}

\strong{Probability mass function (p.m.f.):}
\deqn{f(y) = \frac{1}{n} \sum_{i=1}^{N} (x_i = y)}{f(x) = 1 / N * \sum(x == y)}

\strong{Cummulative distribution function (c.d.f.):}
\deqn{F(y) = \frac{1}{N} \sum_{i=1}^N \mathbf{I}(x_i \leq y)}{F(y) = 1 / N * sum(x <= y)}

\strong{Moment generating functions}:
\itemize{
\item Mean/expectation: \deqn{\bar{x} = \frac{1}{N} \sum_{i=1}^{N} x_i}{mean(x) = 1 / N * sum(x)}
\item Variance: \deqn{\text{VAR}(x) = \frac{1}{N - 1} \sum_{i=1}^{N} (x_i - \bar{x})}{variance(x) = 1 / (N - 1) * sum((x - mean(x))^2)}
}

Third and fourth order moments are also available via \code{\link[=skewness]{skewness()}} and \code{\link[=kurtosis]{kurtosis()}}. For both
different measures (types) are available as defined below. For details see Joanes and Gill (1998).
\itemize{
\item Skewness:
\itemize{
\item Type 1: \deqn{S_1 = \sqrt{N} \frac{\sum_{i=1}^N (x_i - \bar{x})^3}{\sqrt{\big(\sum_{i=1}^N (x_i - \bar{x})^2\big)^3}}}{S1 = sqrt(N) * sum(x - mean(x))^3) / sqrt(sum(x - mean(x))^2)^3}
\item Type 2 (only defined for three or more finite values): \deqn{S_2 = \frac{\sqrt{N \cdot (N - 1)}}{(N - 2)} S_1}{S2 = sqrt(N * (N - 1)) / (N - 2) * S1}
\item Type 3 (default): \deqn{S_3 = \sqrt{(1 - \frac{1}{N})^3} \cdot S_1}{S3 = sqrt((1 - 1 / N)^3) * S1}
\item For more details about the different types of sample skewness see Joanes and Gill (1998).
}
\item Kurtosis:
\itemize{
\item Type 1: \deqn{K_1 = N \cdot \frac{\sum_{i=1}^N (x_i - \bar{x})^4}{\big(\sum_{i=1}^N (x_i - \bar{x})^2\big)^2} - 3}{K1 = N * (sum(x - mean(x))^4) / (sum(x - mean(x))^2)^2 - 3}
\item Type 2 (only defined for four or more finite values): \deqn{K_2 = \frac{((N + 1) \cdot K_1 + 6) \cdot (N - 1)}{(N - 2) \cdot (N - 3)}}{K2 = ((N + 1) * K_1 + 6) * (N - 1)) / ((N - 2) * (N - 3))}
\item Type 3 (default): \deqn{K_3 = \big(1 - \frac{1}{N}\big)^2 \cdot (K_1 + 3) - 3}{K3 = (1 - 1 / N)^2 * (K1 + 3) - 3}
}
}

This function returns the same values that you get from a Z-table. Note
\code{quantile()} is the inverse of \code{cdf()}. Please see the documentation of
\code{\link[=Empirical]{Empirical()}} for some properties of the Empirical distribution, as well as
extensive examples showing to how calculate p-values and confidence
intervals.
}
\examples{

set.seed(28)

X <- Empirical(rnorm(50))
X

mean(X)
variance(X)
skewness(X)
kurtosis(X)

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

### example: allowed types/classes of input arguments

## Single vector (will be coerced to numeric)
Y1  <- rnorm(3, mean = -10)
d1 <- Empirical(Y1)
d1
mean(d1)

## Unnamed list of vectors
Y2 <- list(as.character(rnorm(3, mean = -10)),
           runif(6),
           rpois(4, lambda = 15))
d2 <- Empirical(Y2)
d2
mean(d2)

## Named list of vectors
Y3 <- list("Normal"  = as.character(rnorm(3, mean = -10)),
           "Uniform" = runif(6),
           "Poisson" = rpois(4, lambda = 15))
d3 <- Empirical(Y3)
d3
mean(d3)

## Matrix
Y4 <- matrix(rnorm(20), ncol = 5, dimnames = list(sprintf("D_\%d", 1:4), sprintf("obs_\%d", 1:5)))
d4 <- Empirical(Y4)
d4

## Data frame
d5 <- Empirical(as.data.frame(t(Y4)))
d5

mean(d5)
variance(d5)
skewness(d5)
kurtosis(d5)

pdf(d5, c(-0.5, 0, 0.5, 1)) # Defaults to elementwise = TRUE
pdf(d5, c(-0.5, 0, 0.5, 1), elementwise = FALSE)

cdf(d5, c(-0.5, 0, 0.5, 1)) # Defaults to elementwise = TRUE
cdf(d5, c(-0.5, 0, 0.5, 1), elementwise = FALSE)

quantile(d5, c(0.2, 0.4, 0.6, 0.8)) # Defaults to elementwise = TRUE
quantile(d5, c(0.2, 0.4, 0.6, 0.8), elementwise = FALSE)

}
\references{
Joanes DN and Gill CA (1998). \dQuote{Comparing Measures of
Sample Skewness and Kurtosis.} \emph{Journal of the Royal Statistical
Society D}, \bold{47}(1), 183--189. \doi{10.1111/1467-9884.00122}
}
\seealso{
Other Empirical distribution:
\code{\link[=cdf.Empirical]{cdf.Empirical()}},
\code{\link[=dempirical]{dempirical()}},
\code{\link[=is_continuous.Empirical]{is_continuous.Empirical()}}

Other Normal distribution:
\code{\link[=cdf.Normal]{cdf.Normal()}},
\code{\link[=fit_mle.Normal]{fit_mle.Normal()}},
\code{\link[=pdf.Normal]{pdf.Normal()}},
\code{\link[=quantile.Normal]{quantile.Normal()}}
}
\concept{Empirical distribution}
\concept{Normal distribution}
