% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{Empirical}
\alias{Empirical}
\alias{mean.Empirical}
\alias{variance.Empirical}
\alias{skewness.Empirical}
\alias{kurtosis.Empirical}
\title{Create an Empirical distribution}
\usage{
Empirical(x)

\method{mean}{Empirical}(x, ...)

\method{variance}{Empirical}(x, ...)

\method{skewness}{Empirical}(x, type = 3L, ...)

\method{kurtosis}{Empirical}(x, type = 3L, ...)
}
\arguments{
\item{x}{an object of class \code{Empirical} (see \code{\link[=Empirical]{Empirical()}}).}

\item{...}{currently unused.}

\item{type}{integer between \code{1L} and \code{3L} (default) selecting one
of three algorithms used for calculating sample skewness/kurtosis.
See section Details for more information.}
}
\value{
An \code{Empirical} object.
}
\description{
An empirical distribution consists of a random sample \code{x}.
}
\details{
The constructor function \code{\link[=Empirical]{Empirical()}} allows for a variety of different
objects as main input \code{x}.
\itemize{
\item Vector: Assumes that the vector contains a series of observations from one
empirical distribution.
\item List (named or unnamed) of vectors: Each element in the list describes one
empirical distribution defined by the numeric values in each of the vectors.
\item Matrix/data frame: Each row corresponds to one empirical distribution, whilst
the columns contain the individual observations.
}

Missing values are allowed, however, each distribution requires at least two
finite observations (\code{-Inf}/\code{Inf} is replaced by \code{NA}).
Certain types of moments (see \code{\link[=skewness.Empirical]{skewness.Empirical()}}, \code{\link[=kurtosis.Empirical]{kurtosis.Empirical()}})
require at least three or four finite observations.

\strong{Support}: Set of unique observations in \code{x}.

\strong{Probability mass function (p.m.f.):}
\deqn{f(y) = \frac{1}{n} \sum_{i=1}^{N} (x_i = y)}{f(x) = 1 / N * \sum(x == y)}

\strong{Cummulative distribution function (c.d.f.):}
\deqn{F(y) = \frac{1}{N} \sum_{i=1}^N \mathbf{I}(x_i \leq y)}{F(y) = 1 / N * sum(x <= y)}

\strong{Moment generating functions}:
\itemize{
\item Mean/expectation: \deqn{E(x) = \bar{x} = \frac{1}{N} \sum_{i=1}^{N} x_i}{mean(x) = 1 / N * sum(x)}
\item Variance: \deqn{\text{VAR}(x) = \frac{1}{N - 1} \sum_{i=1}^{N} (x_i - \bar{x})}{variance(x) = 1 / (N - 1) * sum((x - mean(x))^2)}
}

Third and fourth central moments are also available via \code{\link[=skewness]{skewness()}} and \code{\link[=kurtosis]{kurtosis()}}. For both
different types are available as defined below. For details see Joanes and Gill (1998).
\itemize{
\item Skewness:
\itemize{
\item Type 1: \deqn{S_1 = \sqrt{N} \frac{\sum_{i=1}^N (x_i - \bar{x})^3}{\sqrt{\big(\sum_{i=1}^N (x_i - \bar{x})^2\big)^3}}}{S1 = sqrt(N) * sum(x - mean(x))^3) / sqrt(sum(x - mean(x))^2)^3}
\item Type 2 (only defined for three or more finite values): \deqn{S_2 = \frac{\sqrt{N \cdot (N - 1)}}{(N - 2)} S_1}{S2 = sqrt(N * (N - 1)) / (N - 2) * S1}
\item Type 3 (default): \deqn{S_3 = \sqrt{(1 - \frac{1}{N})^3} \cdot S_1}{S3 = sqrt((1 - 1 / N)^3) * S1}
}
\item Kurtosis:
\itemize{
\item Type 1: \deqn{K_1 = N \cdot \frac{\sum_{i=1}^N (x_i - \bar{x})^4}{\big(\sum_{i=1}^N (x_i - \bar{x})^2\big)^2} - 3}{K1 = N * (sum(x - mean(x))^4) / (sum(x - mean(x))^2)^2 - 3}
\item Type 2 (only defined for four or more finite values): \deqn{K_2 = \frac{((N + 1) \cdot K_1 + 6) \cdot (N - 1)}{(N - 2) \cdot (N - 3)}}{K2 = ((N + 1) * K_1 + 6) * (N - 1)) / ((N - 2) * (N - 3))}
\item Type 3 (default): \deqn{K_3 = \big(1 - \frac{1}{N}\big)^2 \cdot (K_1 + 3) - 3}{K3 = (1 - 1 / N)^2 * (K1 + 3) - 3}
}
}
}
\examples{

set.seed(28)

X <- Empirical(rnorm(50))
X

mean(X)
variance(X)
skewness(X)
kurtosis(X)

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

### example: allowed types/classes of input arguments

## Single vector (will be coerced to numeric)
Y1 <- rnorm(3, mean = -10)
d1 <- Empirical(Y1)
d1
mean(d1)

## Unnamed list of vectors
Y2 <- list(as.character(rnorm(3, mean = -10)),
           runif(6),
           rpois(4, lambda = 15))
d2 <- Empirical(Y2)
d2
mean(d2)

## Named list of vectors
Y3 <- list("Normal"  = as.character(rnorm(3, mean = -10)),
           "Uniform" = runif(6),
           "Poisson" = rpois(4, lambda = 15))
d3 <- Empirical(Y3)
d3
mean(d3)

## Matrix
Y4 <- matrix(rnorm(20), ncol = 5,
             dimnames = list(paste0("D_", 1:4), paste0("obs_", 1:5)))
d4 <- Empirical(Y4)
d4

## Data frame
d5 <- Empirical(as.data.frame(Y4))
d5

identical(d4, d5)

mean(d5)
variance(d5)
skewness(d5)
kurtosis(d5)

pdf(d5, c(-0.5, 0, 0.5, 1)) # Defaults to elementwise = TRUE
pdf(d5, c(-0.5, 0, 0.5, 1), elementwise = FALSE)

cdf(d5, c(-0.5, 0, 0.5, 1)) # Defaults to elementwise = TRUE
cdf(d5, c(-0.5, 0, 0.5, 1), elementwise = FALSE)

quantile(d5, c(0.2, 0.4, 0.6, 0.8)) # Defaults to elementwise = TRUE
quantile(d5, c(0.2, 0.4, 0.6, 0.8), elementwise = FALSE)

## The quantile function is the inverse of the distribution
## function (cdf) if x in Y
set.seed(6020)
Y <- round(rlnorm(20, log(3), log(2)), 1)
d <- Empirical(Y)

cdf(d, 4.0)
quantile(d, cdf(d, 4.0))

}
\references{
Joanes DN and Gill CA (1998). \dQuote{Comparing Measures of
Sample Skewness and Kurtosis.} \emph{Journal of the Royal Statistical
Society D}, \bold{47}(1), 183--189. \doi{10.1111/1467-9884.00122}
}
\seealso{
Other Empirical distribution:
\code{\link[=cdf.Empirical]{cdf.Empirical()}},
\code{\link[=dempirical]{dempirical()}},
\code{\link[=pdf.Empirical]{pdf.Empirical()}},
\code{\link[=quantile.Empirical]{quantile.Empirical()}},
\code{\link[=random.Empirical]{random.Empirical()}},
\code{\link[=support.Empirical]{support.Empirical()}}
}
\concept{Empirical distribution}
