% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{Erlang}
\alias{Erlang}
\title{Create an Erlang distribution}
\usage{
Erlang(k = numeric(), lambda = numeric())
}
\arguments{
\item{k}{The shape parameter. Can be any positive integer number.}

\item{lambda}{The rate parameter. Can be any positive number.}
}
\value{
An \code{Erlang} object.
}
\description{
The Erlang distribution is a two-parameter family of continuous probability
distributions with support \eqn{x \in [0,\infty)}.
The two parameters are a positive integer shape parameter \eqn{k} and a
positive real rate parameter \eqn{\lambda}.
The Erlang distribution with shape parameter \eqn{k = 1} simplifies to the
exponential distribution, and it is a special case of the gamma distribution.
It corresponds to a sum of \eqn{k} independent exponential variables with mean
\eqn{1 / \lambda} each.
}
\examples{

set.seed(27)

X <- Erlang(5, 2)
X

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 7))
}
\seealso{
Other continuous distributions:
\code{\link[=Beta]{Beta()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=ChiSquare]{ChiSquare()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=FisherF]{FisherF()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=GEV]{GEV()}},
\code{\link[=GP]{GP()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=RevWeibull]{RevWeibull()}},
\code{\link[=StudentsT]{StudentsT()}},
\code{\link[=Tukey]{Tukey()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Weibull]{Weibull()}}
}
\concept{continuous distributions}
