% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{exponential}
\alias{exponential}
\title{Create a exponential distribution}
\usage{
exponential(rate = 1)
}
\arguments{
\item{rate}{The rate parameter. Can be any positive number. Defaults
to \code{1}.}
}
\value{
A \code{exponential} object.
}
\description{
Create a exponential distribution
}
\examples{

e <- exponential(5)
e

random(e, 10)
pdf(e, 2)
cdf(e, 4)
quantile(e, 0.7)

cdf(e, quantile(e, 0.7))
quantile(e, cdf(e, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{chi_square}},
  \code{\link{fisher_f}}, \code{\link{gamma}},
  \code{\link{normal}}, \code{\link{poisson}},
  \code{\link{students_t}}
}
\concept{continuous distributions}
