% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernoulli.R
\name{fit_mle.Bernoulli}
\alias{fit_mle.Bernoulli}
\title{Fit a Bernoulli distribution to data}
\usage{
\method{fit_mle}{Bernoulli}(d, x, ...)
}
\arguments{
\item{d}{A \code{Bernoulli} object.}

\item{x}{A vector of zeroes and ones.}

\item{...}{Unused.}
}
\value{
a \code{Bernoulli} object
}
\description{
Fit a Bernoulli distribution to data
}
