% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{fit_mle.Exponential}
\alias{fit_mle.Exponential}
\title{Fit an Exponential distribution to data}
\usage{
\method{fit_mle}{Exponential}(d, x, ...)
}
\arguments{
\item{d}{An \code{Exponential} object created by a call to \code{\link[=Exponential]{Exponential()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
An \code{Exponential} object.
}
\description{
Fit an Exponential distribution to data
}
\concept{Exponential distribution}
