% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{fit_mle.Gamma}
\alias{fit_mle.Gamma}
\title{Fit a Gamma distribution to data}
\usage{
\method{fit_mle}{Gamma}(d, x, ...)
}
\arguments{
\item{d}{A \code{Gamma} object created by a call to \code{\link[=Gamma]{Gamma()}}.}

\item{x}{A vector to fit the Gamma distribution to.}

\item{...}{Unused.}
}
\value{
a \code{Gamma} object
}
\description{
Fit a Gamma distribution to data
}
