% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{fit_mle.normal}
\alias{fit_mle.normal}
\title{Fit a normal distribution to data}
\usage{
\method{fit_mle}{normal}(d, x, ...)
}
\arguments{
\item{d}{A \code{normal} object created by a call to \code{\link[=normal]{normal()}}.}

\item{x}{A vector to fit the distribution to.}
}
\value{
A \code{normal} object.
}
\description{
Fit a normal distribution to data
}
\seealso{
Other normal distribution: \code{\link{cdf.normal}},
  \code{\link{pdf.normal}}, \code{\link{quantile.normal}},
  \code{\link{random.normal}}
}
\concept{normal distribution}
