% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{fit_mle.Normal}
\alias{fit_mle.Normal}
\title{Fit a Normal distribution to data}
\usage{
\method{fit_mle}{Normal}(d, x, ...)
}
\arguments{
\item{d}{A \code{Normal} object created by a call to \code{\link[=Normal]{Normal()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A \code{Normal} object.
}
\description{
Fit a Normal distribution to data
}
\seealso{
Other Normal distribution: 
\code{\link{cdf.Normal}()},
\code{\link{pdf.Normal}()},
\code{\link{quantile.Normal}()}
}
\concept{Normal distribution}
