% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\name{fit_mle.Poisson}
\alias{fit_mle.Poisson}
\title{Fit an Poisson distribution to data}
\usage{
\method{fit_mle}{Poisson}(d, x, ...)
}
\arguments{
\item{d}{An \code{Poisson} object created by a call to \code{\link[=Poisson]{Poisson()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
An \code{Poisson} object.
}
\description{
Fit an Poisson distribution to data
}
\concept{Poisson distribution}
