% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.R
\name{geometric}
\alias{geometric}
\title{Create a geometric distribution}
\usage{
geometric(p = 0.5)
}
\arguments{
\item{p}{The success probability for the distribution. \code{p} can be
any value in \link{0, 1}, and defaults to \code{0.5}.}
}
\value{
A \code{geometric} object.
}
\description{
The geometric distribution can be thought of as a generalization
of the \code{\link[=bernoulli]{bernoulli()}} distribution where ask: "if I keep flipping a
coin with probability \code{p} of heads, what is the probability I need
\eqn{k} flips before I get my first heads?" The geometric
distribution is a special case of negative binomial distribution
with parameters TODO.
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a geometric random variable with
success probability \code{p} = \eqn{p}.

TODO: multiple parameterizations BLEH

\strong{Support}: TODO

\strong{Mean}: TODO

\strong{Variance}: TODO

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:

TODO
}
\examples{

g <- geometric(0.3)
g

random(g, 10)
pdf(g, 2)
log_pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

}
\seealso{
Other discrete distributions: \code{\link{bernoulli}},
  \code{\link{binomial}}, \code{\link{log_normal}},
  \code{\link{negative_binomial}}
}
\concept{discrete distributions}
