% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasS3method}
\alias{hasS3method}
\title{Helper functions to check available support for some S3methods}
\usage{
hasS3method(method, classes)
}
\arguments{
\item{method}{character, name of the method (e.g., is_continuous, print, ...)}

\item{classes}{character vector of length > 0, classes to check.}
}
\value{
Returns \code{TRUE} if the method exists for one of the given classes, else \code{FALSE}.
}
\description{
Helper functions to check available support for some S3methods
}
