% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperGeometric.R
\name{HyperGeometric}
\alias{HyperGeometric}
\title{Create a HyperGeometric distribution}
\usage{
HyperGeometric(m, n, k)
}
\arguments{
\item{m}{TODO}

\item{n}{TODO}

\item{k}{TODO}
}
\value{
A \code{HyperGeometric} object.
}
\description{
TODO
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a HyperGeometric random variable with
success probability \code{p} = \eqn{p}.

\strong{Support}: TODO

\strong{Mean}: TODO

\strong{Variance}: TODO

\strong{Probability mass function (p.m.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:

TODO
}
\examples{

X <- HyperGeometric(4, 5, 8)
X

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

}
\seealso{
Other discrete distributions: \code{\link{Bernoulli}},
  \code{\link{Binomial}}, \code{\link{Custom}},
  \code{\link{Geometric}}, \code{\link{Multinomial}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}}
}
\concept{discrete distributions}
