% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{LogNormal}
\alias{LogNormal}
\title{Create a LogNormal distribution}
\usage{
LogNormal(log_mu = 0, log_sigma = 1)
}
\arguments{
\item{log_mu}{TODO explain what the hell this is}

\item{log_sigma}{TODO explain what the hell this is}
}
\value{
A \code{LogNormal} object.
}
\description{
Create a LogNormal distribution
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a LogNormal random variable with
success probability \code{p} = \eqn{p}.

TODO: multiple parameterizations BLEH

\strong{Support}: TODO

\strong{Mean}: TODO

\strong{Variance}: TODO

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:

TODO
}
\examples{

X <- LogNormal(0.3, 2)
X

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

}
\seealso{
Other discrete distributions: \code{\link{Bernoulli}},
  \code{\link{Binomial}}, \code{\link{Geometric}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}}
}
\concept{discrete distributions}
