% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal}
\alias{normal}
\title{Create a normal distribution}
\usage{
normal(mu = 0, sigma = 1)
}
\arguments{
\item{mu}{The location parameter, which is also the mean of the
distribution. Can be any real number. Defaults to \code{0}.}

\item{sigma}{The scale parameter, which is also the standard
deviation of the distribution. Can be any positive number.
Defaults to \code{1}.}
}
\value{
A \code{normal} object.
}
\description{
Create a normal distribution
}
\examples{

n <- normal(5, 2)
n

random(n, 10)
pdf(n, 2)
cdf(n, 4)
quantile(n, 0.7)

cdf(n, quantile(n, 0.7))
quantile(n, cdf(n, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{chi_square}},
  \code{\link{exponential}}, \code{\link{fisher_f}},
  \code{\link{gamma}}, \code{\link{poisson}}
}
\concept{continuous distributions}
