% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{pdf.gamma}
\alias{pdf.gamma}
\title{Evaluate the probability mass function of a gamma distribution}
\usage{
\method{pdf}{gamma}(d, x, ...)
}
\arguments{
\item{d}{A \code{gamma} object created by a call to \code{\link[=gamma]{gamma()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a gamma distribution
}
\examples{

g <- gamma(5, 2)
g

random(g, 10)
pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

cdf(g, quantile(g, 0.7))
quantile(g, cdf(g, 7))

}
