% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralisedPareto.R
\name{pdf.GP}
\alias{pdf.GP}
\alias{log_pdf.GP}
\title{Evaluate the probability mass function of a GP distribution}
\usage{
\method{pdf}{GP}(d, x, ...)

\method{log_pdf}{GP}(d, x, ...)
}
\arguments{
\item{d}{A \code{GP} object created by a call to \code{\link[=GP]{GP()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a GP distribution
}
\examples{

set.seed(27)

X <- GP(0, 2, 0.1)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
