% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{pdf.normal}
\alias{pdf.normal}
\title{Evaluate the probability mass function of a normal distribution}
\usage{
\method{pdf}{normal}(d, x, ...)
}
\arguments{
\item{d}{A \code{normal} object created by a call to \code{\link[=normal]{normal()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
Evaluate the probability mass function of a normal distribution
}
\examples{

n <- normal(5, 2)
n

random(n, 10)
pdf(n, 2)
cdf(n, 4)
quantile(n, 0.7)

cdf(n, quantile(n, 0.7))
quantile(n, cdf(n, 7))

}
