% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pdf}
\alias{pdf}
\alias{pmf}
\title{Evaluate the probability density of a probability distribution}
\usage{
pdf(d, x, ...)

pmf(d, x, ...)
}
\arguments{
\item{d}{A probability distribution object such as those created by
a call to \code{\link[=bernoulli]{bernoulli()}}, \code{\link[=beta]{beta()}}, or \code{\link[=binomial]{binomial()}}.}

\item{x}{A vector of elements whose probabilities you would like to
determine given the distribution \code{d}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of \code{x}.
}
\description{
For discrete distributions, the probabilty mass function. \code{pmf()}
is an alias.
}
