% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson}
\alias{poisson}
\title{Create a poisson distribution}
\usage{
poisson(lambda)
}
\arguments{
\item{lambda}{The shape parameter, which is also the mean and the
variance of the distribution. Can be any positive number.}
}
\value{
A \code{poisson} object.
}
\description{
Poisson distributions are frequently used to model counts.
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the
math will render nicely.

In the following, let \eqn{X} be a Poisson random variable with parameter
\code{lamdba} = \eqn{\lambda}.

\strong{Support}: \eqn{0, 1, 2, 3, ...}

\strong{Mean}: \eqn{\lambda}

\strong{Variance}: \eqn{\lambda}

\strong{Probability mass function (p.m.f)}:

\deqn{P(X = k) = \frac{\lambda^k e^{-\lambda}}{k!}}

\strong{Cumulative distribution function (c.d.f)}:

\deqn{P(X \le k) = e^{-\lambda} \sum_{i = 0}^{\lfloor k \rfloor} \frac{\lambda^i}{i!}}

\strong{Moment generating function (m.g.f)}:

\deqn{\mathbb{E}(e^{tX}) = e^{\lambda (e^t - 1)}}
}
\examples{

p <- poisson(2)
p

random(p, 10)
pdf(p, 2)
cdf(p, 4)
quantile(p, 0.7)

cdf(p, quantile(p, 0.7))
quantile(p, cdf(p, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{chi_square}},
  \code{\link{exponential}}, \code{\link{fisher_f}},
  \code{\link{gamma}}, \code{\link{normal}}
}
\concept{continuous distributions}
