% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernoulli.R
\name{quantile.Bernoulli}
\alias{quantile.Bernoulli}
\title{Determine quantiles of a Bernoulli distribution}
\usage{
\method{quantile}{Bernoulli}(d, p, ...)
}
\arguments{
\item{d}{A \code{Bernoulli} object created by a call to \code{\link[=Bernoulli]{Bernoulli()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

X <- Bernoulli(0.7)
X

random(X, 10)
pdf(X, 1)
log_pdf(X, 1)
cdf(X, 0)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))

}
