% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{quantile.beta}
\alias{quantile.beta}
\title{Determine quantiles of a beta distribution}
\usage{
\method{quantile}{beta}(d, p)
}
\arguments{
\item{d}{A `beta` object created by a call to [beta()].}

\item{p}{A vector of probabilites.}
}
\value{
A vector of quantiles, one for each element of `p`.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

b <- beta(1, 2)
b

random(b, 10)
pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
