% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{quantile.binomial}
\alias{quantile.binomial}
\title{Determine quantiles of a binomial distribution}
\usage{
\method{quantile}{binomial}(d, p, ...)
}
\arguments{
\item{d}{A \code{binomial} object created by a call to \code{\link[=binomial]{binomial()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

b <- binomial(10, 0.2)
b

random(b, 10)
pdf(b, 2L)
log_pdf(b, 2L)
cdf(b, 4L)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 7))

}
