% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.R
\name{quantile.Binomial}
\alias{quantile.Binomial}
\title{Determine quantiles of a Binomial distribution}
\usage{
\method{quantile}{Binomial}(x, probs, ...)
}
\arguments{
\item{x}{A \code{Binomial} object created by a call to \code{\link[=Binomial]{Binomial()}}.}

\item{probs}{A vector of probabilities.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{probs}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

set.seed(27)

X <- Binomial(10, 0.2)
X

mean(X)
variance(X)
skewness(X)
kurtosis(X)

random(X, 10)

pdf(X, 2L)
log_pdf(X, 2L)

cdf(X, 4L)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 7))
}
