% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile.default}
\alias{quantile.default}
\title{Use stats::quantile if passing non-distribution object}
\usage{
\method{quantile}{default}(d, p, names = FALSE, ...)
}
\description{
This function allows us to use the stats::quantile function along side
distributions3::quantile. It accepts both the distributions3 style arguments
(d, p) and stats style arguments (x, probs) as well as any combination of
the two.
}
