% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.R
\name{quantile.geometric}
\alias{quantile.geometric}
\title{Determine quantiles of a geometric distribution}
\usage{
\method{quantile}{geometric}(d, p, ...)
}
\arguments{
\item{d}{A \code{geometric} object created by a call to \code{\link[=geometric]{geometric()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
Determine quantiles of a geometric distribution
}
\examples{

g <- geometric(0.3)
g

random(g, 10)
pdf(g, 2)
log_pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

}
\seealso{
Other geometric distribution: \code{\link{cdf.geometric}},
  \code{\link{pdf.geometric}},
  \code{\link{random.geometric}}
}
\concept{geometric distribution}
