% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralisedPareto.R
\name{quantile.GP}
\alias{quantile.GP}
\title{Determine quantiles of a GP distribution}
\usage{
\method{quantile}{GP}(x, probs, ...)
}
\arguments{
\item{x}{A \code{GP} object created by a call to \code{\link[=GP]{GP()}}.}

\item{probs}{A vector of probabilities.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{probs}.
}
\description{
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

set.seed(27)

X <- GP(0, 2, 0.1)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
