% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{quantile.normal}
\alias{quantile.normal}
\title{Determine quantiles of a normal distribution}
\usage{
\method{quantile}{normal}(d, p, ...)
}
\arguments{
\item{d}{A \code{normal} object created by a call to \code{\link[=normal]{normal()}}.}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of \code{p}.
}
\description{
This function returns the same values that you get from a Z-table. Note
\code{quantile()} is the inverse of \code{cdf()}.
}
\examples{

# TODO: ggplot example illustrating that you always get the lower
# tail probability

}
