% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{random.binomial}
\alias{random.binomial}
\title{Draw a random sample from a binomial distribution}
\usage{
\method{random}{binomial}(d, n = 1L, ...)
}
\arguments{
\item{d}{A `binomial` object created by a call to [binomial()].}

\item{n}{The number of samples to draw. Defaults to `1L`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
An integer vector containing values between `0` and `d$size`
  of length `n`.
}
\description{
Draw a random sample from a binomial distribution
}
\examples{

b <- binomial(10, 0.2)
b

random(b, 10)
pdf(b, 2L)
cdf(b, 4L)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 7))

}
