% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{random.Exponential}
\alias{random.Exponential}
\title{Draw a random sample from an Exponential distribution}
\usage{
\method{random}{Exponential}(x, n = 1L, drop = TRUE, ...)
}
\arguments{
\item{x}{An \code{Exponential} object created by a call to \code{\link[=Exponential]{Exponential()}}.}

\item{n}{The number of samples to draw. Defaults to \code{1L}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
In case of a single distribution object or \code{n = 1}, either a numeric
vector of length \code{n} (if \code{drop = TRUE}, default) or a \code{matrix} with \code{n} columns
(if \code{drop = FALSE}).
}
\description{
Draw a random sample from an Exponential distribution
}
\examples{

set.seed(27)

X <- Exponential(5)
X

mean(X)
variance(X)
skewness(X)
kurtosis(X)

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 7))
}
