% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{random}
\alias{random}
\title{Draw a random sample from a probability distribution}
\usage{
random(x, n = 1L, ...)
}
\arguments{
\item{x}{A probability distribution object such as those created by
a call to \code{\link[=Bernoulli]{Bernoulli()}}, \code{\link[=Beta]{Beta()}}, or \code{\link[=Binomial]{Binomial()}}.}

\item{n}{The number of samples to draw. Should be a positive
integer. Defaults to \code{1L}.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\description{
Draw a random sample from a probability distribution
}
\examples{

X <- Normal()

random(X, 10)
}
