% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernoulli.R, R/Binomial.R, R/Normal.R,
%   R/Poisson.R, R/SinhArcsinh.R, R/Uniform.R, R/score-hessian.R
\name{score.Bernoulli}
\alias{score.Bernoulli}
\alias{hessian.Bernoulli}
\alias{score.Binomial}
\alias{hessian.Binomial}
\alias{score.Normal}
\alias{hessian.Normal}
\alias{score.Poisson}
\alias{hessian.Poisson}
\alias{score.SinhArcsinh}
\alias{hessian.SinhArcsinh}
\alias{score.Uniform}
\alias{hessian.Uniform}
\alias{score}
\alias{hessian}
\alias{score.distribution}
\alias{hessian.distribution}
\title{Generic functions and methods for computing score and Hessian}
\usage{
score(d, ...)

hessian(d, ...)

\method{score}{distribution}(d, x, which = NULL, drop = TRUE, eps = .Machine$double.eps^(1/3), ...)

\method{hessian}{distribution}(
  d,
  x,
  which = NULL,
  drop = TRUE,
  expected = FALSE,
  eps = .Machine$double.eps^(1/4),
  ...
)
}
\arguments{
\item{d}{An object, typically a distribution object, e.g., as created by
\code{\link[=Normal]{Normal()}} or \code{\link[=Binomial]{Binomial()}}.}

\item{x}{A vector of elements whose score/Hessian should be determined given the
distribution \code{d}. Either \code{d} and \code{x} need to have the same length or length 1.}

\item{which}{character or \code{NULL} (default). Character labels for the derivatives
to be included in the score or Hessian respectively. In \code{score} the possible
values are (combinations of) the parameter names (e.g., \code{"mu"} and/or \code{"sigma"}).
In \code{hessian} additionally the cross-derivatives (e.g., \code{mu:sigma} and \code{sigma:mu})
are available. By default (if \code{which = NULL}) all elements of the score or Hessian
should be computed.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Arguments passed to methods.}

\item{expected}{logical. Should the expected Hessian be computed? If \code{FALSE} the
observed Hessian is computed. Some methods might only support only one or the
the other option and defaults might differ.}

\item{eps}{numeric. Tolerance when obtaining the score or Hessian via numeric
differentiation.}
}
\value{
Either a numeric matrix with suitable column names is returned or a numeric
vector (provided that \code{which} has length 1 and \code{drop = TRUE}).
}
\description{
The generic functions \code{score} and \code{hessian} along with the corresponding
methods for distribution objects enable the computation of the
score (first derivative of the log-likelihood with respect to the
parameters) and Hessian (corresponding second derivative).
}
\details{
In the methods for dedicated distributions analytical results for computing
the score or Hessian should be used. All \code{hessian} methods should implemented
the \code{expected} argument with an appropriate default. If a \code{hessian} method only
supports the expected or the observed Hessian, then an error should be issued
if needed.

In the fallback methods for general \code{distribution} objects a simple differencing
approach is used. Either differences of \code{log_pdf} (for the \code{score}) or of
\code{score} (for the \code{hessian}) with slightly modified parameters are used. If
the parameters in a distribution are on the boundary of the parameter space
this might lead to errors in the computation.
}
\examples{
X <- Normal(mu = c(0, 1, 2), sigma = c(2, 1, 1))
x <- c(0, 0, 1)
score(X, x)
hessian(X, x)
hessian(X, x, expected = TRUE)

h <- hessian(X[1], x[1], expected = TRUE)
matrix(h, ncol = 2, dimnames = list(c("mu", "sigma"), c("mu", "sigma")))

## Comparison of analytic and numeric score/Hessian (Normal(3, 2))
X <- Normal(mu = 3, sigma = 2)
x <- seq(0, 6, by = 0.01)

#' ## score:   derivative of log-likelihood by parameter sigma
s_analytic <- score(X, x, which = "sigma")
s_numeric  <- distributions3:::score.distribution(X, x, which = "sigma")
message("Sum of absolute differences (score): ", sum(abs(s_analytic - s_numeric)))

matplot(x, cbind(s_analytic, s_numeric), col = 1:2, type = "l", lty = 1:2,
        lwd = 3, xlab = "x", main = "score - analytic vs. numeric solution",
        ylab = expression(partialdiff * l(x) / partialdiff * sigma))
legend("topleft", legend = c("analytic score", "numeric score"),
       bty = "n", pch = NA, lty = 1:2, col = 1:2, lwd = 3)
abline(h = 0, v = 3, lty = 3)

#' ## Hessian: second derivative of log-likelihood by sigma^2
h_analytic <- hessian(X, x, which = "sigma")
h_numeric  <- distributions3:::hessian.distribution(X, x, which = "sigma")
message("Sum of absolute differences (Hessian): ", sum(abs(h_analytic - h_numeric)))

matplot(x, cbind(h_analytic, h_numeric), col = 1:2, type = "l", lty = 1:2,
        lwd = 3, xlab = "x", main = "Hessian - analytic vs. numeric solution",
        ylab = expression(partialdiff^2 * l(x) / partialdiff * sigma^2))
legend("topleft", legend = c("analytic score", "numeric score"),
       bty = "n", pch = NA, lty = 1:2, col = 1:2, lwd = 3)
abline(v = 3, lty = 3)


}
