% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-hessian.R
\name{score}
\alias{score}
\alias{hessian}
\alias{hessian.distribution}
\alias{score.Normal}
\alias{hessian.Normal}
\alias{score.Poisson}
\alias{hessian.Poisson}
\alias{score.Bernoulli}
\alias{hessian.Bernoulli}
\alias{score.Binomial}
\alias{hessian.Binomial}
\title{Generic functions and methods for computing score and Hessian}
\usage{
score(d, ...)

hessian(d, ...)

\method{hessian}{distribution}(
  d,
  x,
  which = NULL,
  drop = TRUE,
  expected = FALSE,
  eps = .Machine$double.eps^(1/4),
  ...
)

\method{score}{Normal}(d, x, which = NULL, drop = TRUE, ...)

\method{hessian}{Normal}(d, x, which = NULL, drop = TRUE, expected = FALSE, ...)

\method{score}{Poisson}(d, x, which = "lambda", drop = TRUE, ...)

\method{hessian}{Poisson}(d, x, which = "lambda", drop = TRUE, expected = FALSE, ...)

\method{score}{Bernoulli}(d, x, which = "p", drop = TRUE, ...)

\method{hessian}{Bernoulli}(d, x, which = "p", drop = TRUE, expected = FALSE, ...)

\method{score}{Binomial}(d, x, which = "p", drop = TRUE, ...)

\method{hessian}{Binomial}(d, x, which = "p", drop = TRUE, expected = FALSE, ...)
}
\arguments{
\item{d}{An object, typically a distribution object., e.g., as created by
\code{\link[distributions3]{Normal}} or \code{\link[distributions3]{Binomial}}.}

\item{...}{Arguments passed to methods}

\item{x}{A vector of elements whose score/Hessian should be determined given the
distribution \code{d}. Either \code{d} and \code{x} need to have the same length or length 1.}

\item{which}{character or \code{NULL} (default). Character labels for the derivatives
to be included in the score or Hessian respectively. In \code{score} the possible
values are (combinations of) the parameter names (e.g., \code{"mu"} and/or \code{"sigma"}).
In \code{hessian} additionally the cross-derivatives (e.g., \code{mu:sigma} and \code{sigma:mu})
are available. By default (if \code{which = NULL}) all elements of the score or Hessian
should be computed.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{expected}{logical. Should the expected Hessian be computed? If \code{FALSE} the
observed Hessian is computed. Some methods might only support only one or the
the other option and defaults might differ.}

\item{eps}{numeric. Tolerance when obtaining the score or Hessian via numeric
differentiation.}
}
\value{
Either a numeric matrix with suitable column names is returned or a numeric
vector provided that \code{which} has length 1 and \code{drop = TRUE} (the default).
}
\description{
The functions \code{score} and \code{hessian} are generic functions along with
methods for distribution objects, enabling the computation of the
score (first derivative of the log-likelihood with respect to the
parameters) and Hessian (corresponding second derivative).
}
\details{
In the methods for dedicated distributions analytical results for computing
the score or Hessian should be used.

In the fallback methods for general \code{distribution} objects a simple differencing
approach is used. Either differences of \code{log_pdf} (for the \code{score}) or of
\code{score} (for the \code{hessian}) with slightly modified parameters are used. If
the parameters in a distribution are on the boundary of the parameter space
this might lead to errors in the computation.
}
\examples{
X <- Normal(mu = c(0, 1, 2), sigma = c(2, 1, 1))
x <- c(0, 0, 1)
score(X, x)
hessian(X, x)
hessian(X, x, expected = TRUE)
}
