% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{dshash}
\alias{dshash}
\alias{pshash}
\alias{qshash}
\alias{rshash}
\title{The Sinh-Arcsinh (SHASH) distribution}
\usage{
dshash(x, mu = 0, sigma = 1, nu = 1, tau = 1, log = FALSE, cores = NULL)

pshash(
  q,
  mu = 0,
  sigma = 1,
  nu = 1,
  tau = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  cores = NULL
)

qshash(
  p,
  mu = 0,
  sigma = 1,
  nu = 1,
  tau = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  cores = NULL
)

rshash(n, mu = 0, sigma = 1, nu = 1, tau = 1, cores = NULL)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{mu, sigma, nu, tau}{vector of (non-negative) SHASH parameters.}

\item{log, log.p}{logical indicating whether probabilities p are given as log(p).}

\item{cores}{integer. Number of cores/threads to be used (requires OMP support).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical indicating whether probabilities are \eqn{P[X \le x]} (lower tail) or \eqn{P[X > x]} (upper tail).}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\description{
Density, distribution function, quantile function, and random
generation for the sinh-arcsinh distribution with four parameters
\code{mu}, \code{sigma}, \code{nu}, and \code{tau}.
}
\details{
The Sinh-Arcsinh generalizes the Normal distribution by separately
controlling location, scale, skewness, and tail-heaviness and can thus produce a
wide range of shapes. Using \code{nu = 1} and \code{tau = 1} results in a
normal distribution.

All functions follow the usual conventions of d/p/q/r functions in base R.
}
\examples{
## TODO(R): Add a couple of examples

}
\concept{SHASH}
\keyword{distribution}
