% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{SHASH}
\alias{SHASH}
\title{Create a Sinh-Arcsinh (SHASH) distribution}
\usage{
SHASH(mu = 0, sigma = 1, nu = 1, tau = 1)
}
\arguments{
\item{mu}{The location parameter, written \eqn{\mu} in textbooks.
Defaults to \code{0}.}

\item{sigma}{The scale parameter, written \eqn{\sigma} in textbooks.
Can be any positive number. Defaults to \code{1}.}

\item{nu}{The skewness parameter \eqn{\nu}, defaults to \code{1}.}

\item{tau}{The kurtosis parameter \eqn{\tau}, defaults to \code{1}.}
}
\value{
A \code{SHASH} object.
}
\description{
The Sinh-Arcsinh (SHASH) distribution is a four-parameter distribution
with support on the real space that generalizes the Normal distribution by separately
controlling location, scale, skewness, and tail-heaviness. It can produce a
wide range of shapes, including symmetric and asymmetric forms and both heavier
and lighter tails, making it useful for modeling real-valued data with
non-normal behavior. Using \eqn{\nu = 1} and \eqn{\tau = 1} results
in a standard normal distribution.
}
\details{
We recommend reading this documentation on
\url{https://zeileis.github.io/distributions3/}, where the math
will render with additional detail and much greater clarity.

TODO(R): List details about the SHASH distribution as implemented.
TODO(R): Check references on ?gamlss.dist::SHASH

In the following, let \eqn{X} be a SHASH random variable with mean
\code{mu} = \eqn{\mu}, \code{sigma} = \eqn{\sigma}, \code{nu} = \eqn{\nu}, and
\code{tau} = \eqn{\tau}.

\strong{Support}: \eqn{R}, the set of all real numbers

\strong{Mean}: ...

\strong{Variance}: ...

\strong{Probability density function (p.d.f)}:

\deqn{
    f(x) = ...
  }{
    f(x) = ...
  }

\strong{Cumulative distribution function (c.d.f)}:

The cumulative distribution function has the form

\deqn{
    F(t) = ...
  }{
    F(t) = ...
  }

\strong{Moment generating function (m.g.f)}:

\deqn{
    ...
  }{
    ...
  }
}
\examples{

## SHASH(), by default, uses nu = 1, tau = 1 which
## results in the standard normal distribution
set.seed(6020)
X <- SHASH() # Uses mu = 1, sigma = 0, nu = 1, tau = 1)
x <- random(X, 300)
qqnorm(x); qqline(x, col = 2, lwd = 2)
c(mean = mean(x), sd = sd(x))

## TODO(R): Write examples
## set.seed(27)

## X <- SHASH(5, 2)
## X

## mean(X)
## variance(X)
## skewness(X)
## kurtosis(X)

## random(X, 10)

## pdf(X, 2)
## log_pdf(X, 2)

## cdf(X, 4)
## quantile(X, 0.7)

## ### example: calculating p-values for two-sided Z-test

## # here the null hypothesis is H_0: mu = 3
## # and we assume sigma = 2

## # exactly the same as: Z <- SHASH(0, 1)
## Z <- SHASH()

## # data to test
## x <- c(3, 7, 11, 0, 7, 0, 4, 5, 6, 2)
## nx <- length(x)

## # calculate the z-statistic
## z_stat <- (mean(x) - 3) / (2 / sqrt(nx))
## z_stat

## # calculate the two-sided p-value
## 1 - cdf(Z, abs(z_stat)) + cdf(Z, -abs(z_stat))

## # exactly equivalent to the above
## 2 * cdf(Z, -abs(z_stat))

## # p-value for one-sided test
## # H_0: mu <= 3   vs   H_A: mu > 3
## 1 - cdf(Z, z_stat)

## # p-value for one-sided test
## # H_0: mu >= 3   vs   H_A: mu < 3
## cdf(Z, z_stat)

## ### example: calculating a 88 percent Z CI for a mean

## # same `x` as before, still assume `sigma = 2`

## # lower-bound
## mean(x) - quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## # upper-bound
## mean(x) + quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## # equivalent to
## mean(x) + c(-1, 1) * quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## # also equivalent to
## mean(x) + quantile(Z, 0.12 / 2) * 2 / sqrt(nx)
## mean(x) + quantile(Z, 1 - 0.12 / 2) * 2 / sqrt(nx)

## ### generating random samples and plugging in ks.test()

## set.seed(27)

## # generate a random sample
## ns <- random(SHASH(3, 7), 26)

## # test if sample is SHASH(3, 7)
## ks.test(ns, pnorm, mean = 3, sd = 7)

## # test if sample is gamma(8, 3) using base R pgamma()
## ks.test(ns, pgamma, shape = 8, rate = 3)

## ### MISC

## # note that the cdf() and quantile() functions are inverses
## cdf(X, quantile(X, 0.7))
## quantile(X, cdf(X, 7))

}
\seealso{
Other continuous distributions:
\code{\link[=Beta]{Beta()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=ChiSquare]{ChiSquare()}},
\code{\link[=Erlang]{Erlang()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=FisherF]{FisherF()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=GEV]{GEV()}},
\code{\link[=GP]{GP()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=RevWeibull]{RevWeibull()}},
\code{\link[=StudentsT]{StudentsT()}},
\code{\link[=Tukey]{Tukey()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Weibull]{Weibull()}}
}
\concept{continuous distributions}
