% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinhArcsinh.R
\name{SinhArcsinh}
\alias{SinhArcsinh}
\title{Create a Sinh-Arcsinh (SHASH) distribution}
\usage{
SinhArcsinh(mu = 0, sigma = 1, nu = 1, tau = 1)
}
\arguments{
\item{mu}{The location parameter, written \eqn{\mu} in textbooks.
Defaults to \code{0}.}

\item{sigma}{The scale parameter, written \eqn{\sigma} in textbooks.
Can be any positive number. Defaults to \code{1}.}

\item{nu}{The skewness parameter \eqn{\nu}, defaults to \code{1}.}

\item{tau}{The kurtosis parameter \eqn{\tau}, defaults to \code{1}.}
}
\value{
A \code{SinhArcsinh} object.
}
\description{
The Sinh-Arcsinh (SHASH) distribution is a four-parameter distribution
with support on the real space that generalizes the Normal distribution by separately
controlling location, scale, skewness, and tail-heaviness. It can produce a
wide range of shapes, including symmetric and asymmetric forms and both heavier
and lighter tails, making it useful for modeling real-valued data with
non-normal behavior. Using \eqn{\nu = 1} and \eqn{\tau = 1} results
in a standard normal distribution.
}
\details{
We recommend reading this documentation on
\url{https://zeileis.github.io/distributions3/}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a Sinh-Arcsinh random variable with mean
\code{mu} = \eqn{\mu}, \code{sigma} = \eqn{\sigma}, \code{nu} = \eqn{\nu}, and
\code{tau} = \eqn{\tau}.

\strong{Support}: \eqn{R}, the set of all real numbers

\strong{Probability density function (p.d.f)}:

\deqn{
    f(t) = \frac{1}{\sigma \sqrt{1 + z^2}} \cdot \frac{1}{2}\left(\tau e^{\tau w} + \nu e^{-\nu w}\right) \cdot \phi\left( \frac{1}{2}\left(e^{\tau w} - e^{-\nu w}\right) \right)
  }{
    f(t) = (1 / (sigma * sqrt(1 + z^2))) * 0.5 * (tau * exp(tau * w) + nu * exp(-nu * w)) * phi(0.5 * (exp(tau * w) - exp(-nu * w)))
  }

\deqn{
    \text{where } z = \frac{t - \mu}{\sigma}, ~w = \operatorname{asinh}(z), ~\text{and } \phi() \text{ is the standard normal PDF.}
  }{
    where z = (x - mu) / sigma, w = asinh(z), and phi() is the standard normal PDF.
  }

\strong{Cumulative distribution function (c.d.f)}:

\deqn{
    F(t) = \Phi\left( H\left(\operatorname{asinh}\left(\frac{t - \mu}{\sigma}\right)\right) \right)
  }{
    F(t) = Phi(H(asinh((y - mu) / sigma)))
  }

\deqn{
    \text{where } H(w) = \frac{1}{2}\left(e^{\tau w} - e^{-\nu w}\right), ~\text{and } \Phi() \text{ the standard normal CDF.}
  }{
    where H(w) = 0.5 * (exp(tau * w) - exp(-nu * w)), and Phi() the standard normal CDF.
  }
}
\examples{

## SinhArcsinh() by default uses nu = 1, tau = 1 which
## results in the standard normal distribution
set.seed(6020)
X <- SinhArcsinh() # Uses mu = 1, sigma = 0, nu = 1, tau = 1)
x <- random(X, 300)
qqnorm(x); qqline(x, col = 2, lwd = 2)
curve(pdf(X, x), xlim = c(-5, 5), main = paste(X, "density"))

## Calculation of central moments is based on numeric integration,
## thus not being identical to the standard normal distribution
c(mean = mean(x), sd = sd(x))

## Skewed Sinh-Arcsinh distribution
X <- SinhArcsinh(mu = 7, sigma = 2, nu = c(0.7, 1, 0.7), tau = c(1, 0.7, 0.7))
as.matrix(X)

## Visualization of density functions using different parameters for nu/tau
curve(pdf(X[1], x), xlim = c(0, 20), ylim = c(0, 0.2), main = "Density function")
curve(pdf(X[2], x), xlim = c(0, 20), col = 2, add = TRUE)
curve(pdf(X[3], x), xlim = c(0, 20), col = 4, add = TRUE)

## Visualization of distribution function using different parameters for nu/tau
curve(cdf(X[1], x), xlim = c(0, 20), ylim = 0:1, main = "Distribution function")
curve(cdf(X[2], x), xlim = c(0, 20), col = 2, add = TRUE)
curve(cdf(X[3], x), xlim = c(0, 20), col = 4, add = TRUE)

## Central moments
mean(X)
variance(X)
skewness(X)
kurtosis(X)

## Drawing random values
random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

# note that the cdf() and quantile() functions are inverses
X <- SinhArcsinh(mu = 3, sigma = 2, nu = 0.9, tau = 1.2)
cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 7))

}
\seealso{
Other continuous distributions:
\code{\link[=Beta]{Beta()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=ChiSquare]{ChiSquare()}},
\code{\link[=Erlang]{Erlang()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=FisherF]{FisherF()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=GEV]{GEV()}},
\code{\link[=GP]{GP()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=RevWeibull]{RevWeibull()}},
\code{\link[=StudentsT]{StudentsT()}},
\code{\link[=Tukey]{Tukey()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Weibull]{Weibull()}}
}
\concept{continuous distributions}
