% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t.R
\name{students_t}
\alias{students_t}
\title{Create a Student's T distribution}
\usage{
students_t(df)
}
\arguments{
\item{df}{Degrees of freedom. Can be any positive number.}
}
\value{
A \code{students_t} object.
}
\description{
TODO
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a students_t random variable with mean
\code{mu} = \eqn{\mu} and standard deviation \code{sigma} = \eqn{\sigma}.

\strong{Support}: \eqn{\mathbb{R}}{R}, the set of all real numbers

\strong{Mean}: Undefined unless \eqn{\nu \ge 2}, in which case the mean is
zero.

\strong{Variance}:

\deqn{
    \frac{\nu}{\nu - 2}
  }

Undefined if \eqn{\nu < 1}, infinite when \eqn{1 < \n \le 2}.

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:
}
\examples{

# TODO

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{chi_square}},
  \code{\link{exponential}}, \code{\link{fisher_f}},
  \code{\link{gamma}}, \code{\link{normal}},
  \code{\link{poisson}}
}
\concept{continuous distributions}
