% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentsT.R
\name{StudentsT}
\alias{StudentsT}
\title{Create a Student's T distribution}
\usage{
StudentsT(df)
}
\arguments{
\item{df}{Degrees of freedom. Can be any positive number.}
}
\value{
A \code{StudentsT} object.
}
\description{
TODO
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a StudentsT random variable with mean
\code{mu} = \eqn{\mu} and standard deviation \code{sigma} = \eqn{\sigma}.

\strong{Support}: \eqn{\mathbb{R}}{R}, the set of all real numbers

\strong{Mean}: Undefined unless \eqn{\nu \ge 2}, in which case the mean is
zero.

\strong{Variance}:

\deqn{
    \frac{\nu}{\nu - 2}
  }

Undefined if \eqn{\nu < 1}, infinite when \eqn{1 < \n \le 2}.

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:
}
\examples{

# TODO

}
\seealso{
Other continuous distributions: \code{\link{Beta}},
  \code{\link{Cauchy}}, \code{\link{ChiSquare}},
  \code{\link{Exponential}}, \code{\link{FisherF}},
  \code{\link{Gamma}}, \code{\link{Normal}},
  \code{\link{Uniform}}, \code{\link{random.Normal}}
}
\concept{continuous distributions}
