% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{suff_stat.bernoulli}
\alias{suff_stat.bernoulli}
\title{Compute the sufficient statistics for a bernoulli distribution from data}
\usage{
\method{suff_stat}{bernoulli}(d, x, ...)
}
\arguments{
\item{d}{A \code{bernoulli} object.}

\item{x}{A vector of zeroes and ones.}
}
\value{
A named list of the sufficient statistics of the bernoulli distribution
\describe{
\item{\code{successes}}{The number of successful trials (\code{x == 1})}
\item{\code{failures}}{The number of failed trials (\code{x == 0})}
}
}
\description{
Compute the sufficient statistics for a bernoulli distribution from data
}
