% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{suff_stat.exponential}
\alias{suff_stat.exponential}
\title{Compute the sufficient statistics of an exponential distribution from data}
\usage{
\method{suff_stat}{exponential}(d, x, ...)
}
\description{
Compute the sufficient statistics of an exponential distribution from data
}
