% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exponential.R
\name{suff_stat.Exponential}
\alias{suff_stat.Exponential}
\title{Compute the sufficient statistics of an Exponential distribution from data}
\usage{
\method{suff_stat}{Exponential}(d, x, ...)
}
\arguments{
\item{d}{An \code{Exponential} object created by a call to \code{\link[=Exponential]{Exponential()}}.}

\item{x}{A vector of data.}
}
\value{
A named list of the sufficient statistics of Exponential distribution
\describe{
\item{\code{sum}}{The sum of the data}
\item{\code{samples}}{The number of samples in the data}
}
}
\description{
Compute the sufficient statistics of an Exponential distribution from data
}
