% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{suff_stat.gamma}
\alias{suff_stat.gamma}
\title{Compute the sufficient statistics for a bernoulli distribution from data}
\usage{
\method{suff_stat}{gamma}(d, x, ...)
}
\value{
A \code{gamma} object.
}
\description{
Several important distributions are special cases of the gamma
distribution. When the shape parameter is \code{1}, the gamma is an
gamma. When the (scale?) parametre is \code{2} the gamma is a
chi square.
}
\examples{

g <- gamma(5, 2)
g

random(g, 10)
pdf(g, 2)
log_pdf(g, 2)
cdf(g, 4)
quantile(g, 0.7)

cdf(g, quantile(g, 0.7))
quantile(g, cdf(g, 7))

}
\seealso{
Other continuous distributions: \code{\link{beta}},
  \code{\link{cauchy}}, \code{\link{chi_square}},
  \code{\link{exponential}}, \code{\link{fisher_f}},
  \code{\link{normal}}, \code{\link{poisson}},
  \code{\link{students_t}}, \code{\link{uniform}}
}
