% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{suff_stat.Gamma}
\alias{suff_stat.Gamma}
\title{Compute the sufficient statistics for a Gamma distribution from data}
\usage{
\method{suff_stat}{Gamma}(d, x, ...)
}
\arguments{
\item{d}{A \code{Gamma} object created by a call to \code{\link[=Gamma]{Gamma()}}.}

\item{x}{A vector to fit the Gamma distribution to.}

\item{...}{Unused.}
}
\value{
a \code{Gamma} object
}
\description{
\itemize{
\item \code{sum}: The sum of the data.
\item \code{log_sum}: The log of the sum of the data.
\item \code{samples}: The number of samples in the data.
}
}
