% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{suff_stat.normal}
\alias{suff_stat.normal}
\title{Compute the sufficient statistics for a normal distribution from data}
\usage{
\method{suff_stat}{normal}(d, x, ...)
}
\description{
Compute the sufficient statistics for a normal distribution from data
}
