% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{suff_stat.Normal}
\alias{suff_stat.Normal}
\title{Compute the sufficient statistics for a Normal distribution from data}
\usage{
\method{suff_stat}{Normal}(d, x, ...)
}
\arguments{
\item{d}{A \code{Normal} object created by a call to \code{\link[=Normal]{Normal()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the normal
distribution:
\itemize{
\item \code{mu}: The sample mean of the data.
\item \code{sigma}: The sample standard deviation of the data.
\item \code{samples}: The number of samples in the data.
}
}
\description{
Compute the sufficient statistics for a Normal distribution from data
}
