% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{suff_stat.poisson}
\alias{suff_stat.poisson}
\title{Compute the sufficient statistics of an poisson distribution from data}
\usage{
\method{suff_stat}{poisson}(d, x, ...)
}
\description{
Compute the sufficient statistics of an poisson distribution from data
}
