% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\name{suff_stat.Poisson}
\alias{suff_stat.Poisson}
\title{Compute the sufficient statistics of an Poisson distribution from data}
\usage{
\method{suff_stat}{Poisson}(d, x, ...)
}
\arguments{
\item{d}{An \code{Poisson} object created by a call to \code{\link[=Poisson]{Poisson()}}.}

\item{x}{A vector of data.}

\item{...}{Unused.}
}
\value{
A named list of the sufficient statistics of the Poisson
distribution:
\itemize{
\item \code{sum}: The sum of the data.
\item \code{samples}: The number of samples in the data.
}
}
\description{
Compute the sufficient statistics of an Poisson distribution from data
}
