% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{suff_stat}
\alias{suff_stat}
\title{Compute the sufficient statistics of a distribution from data}
\usage{
suff_stat(d, x, ...)
}
\value{
a named list of sufficient statistics
}
\description{
Compute the sufficient statistics of a distribution from data
}
