% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{suff_stat}
\alias{suff_stat}
\title{Compute the sufficient statistics of a distribution from data}
\usage{
suff_stat(d, x, ...)
}
\arguments{
\item{d}{A probability distribution object such as those created by
a call to \code{\link[=Bernoulli]{Bernoulli()}}, \code{\link[=Beta]{Beta()}}, or \code{\link[=Binomial]{Binomial()}}.}

\item{x}{A vector of data to compute the likelihood.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
a named list of sufficient statistics
}
\description{
Compute the sufficient statistics of a distribution from data
}
