% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{support.ChiSquare}
\alias{support.ChiSquare}
\title{Return the support of the ChiSquare distribution}
\usage{
\method{support}{ChiSquare}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{ChiSquare} object created by a call to \code{\link[=ChiSquare]{ChiSquare()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the ChiSquare distribution
}
