% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{support.Empirical}
\alias{support.Empirical}
\title{Return the support of the Empirical distribution}
\usage{
\method{support}{Empirical}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{an \code{Empirical} object created by a call to \code{\link[=Empirical]{Empirical()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{currently not used.}
}
\value{
In case of a single distribution object, a numeric vector of length 2
with the minimum and maximum value of the support (if \code{drop = TRUE}, default)
or a \code{matrix} with 2 columns. In case of a vectorized distribution object, a
matrix with 2 columns containing all minima and maxima.
}
\description{
Though the support of an empirical distribution is
defined by its unique numeric values this function
returns only the range, i.e., the lowest (minimum)
and highest (maximum) observation as the outer
bounds of the support.
}
\seealso{
Other Empirical distribution:
\code{\link[=Empirical]{Empirical()}},
\code{\link[=cdf.Empirical]{cdf.Empirical()}},
\code{\link[=dempirical]{dempirical()}},
\code{\link[=pdf.Empirical]{pdf.Empirical()}},
\code{\link[=quantile.Empirical]{quantile.Empirical()}},
\code{\link[=random.Empirical]{random.Empirical()}}
}
\concept{Empirical distribution}
