% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erlang.R
\name{support.Erlang}
\alias{support.Erlang}
\title{Return the support of the Erlang distribution}
\usage{
\method{support}{Erlang}(d)
}
\arguments{
\item{d}{An \code{Erlang} object created by a call to \code{\link[=Erlang]{Erlang()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Erlang distribution
}
