% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{support.LogNormal}
\alias{support.LogNormal}
\title{Return the support of the LogNormal distribution}
\usage{
\method{support}{LogNormal}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{LogNormal} object created by a call to \code{\link[=LogNormal]{LogNormal()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the LogNormal distribution
}
