% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{support.Normal}
\alias{support.Normal}
\title{Return the support of the Normal distribution}
\usage{
\method{support}{Normal}(d)
}
\arguments{
\item{d}{An \code{Normal} object created by a call to \code{\link[=Normal]{Normal()}}.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Normal distribution
}
