% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\name{support.Poisson}
\alias{support.Poisson}
\title{Return the support of the Poisson distribution}
\usage{
\method{support}{Poisson}(d, drop = TRUE, ...)
}
\arguments{
\item{d}{An \code{Poisson} object created by a call to \code{\link[=Poisson]{Poisson()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{...}{Currently not used.}
}
\value{
A vector of length 2 with the minimum and maximum value of the support.
}
\description{
Return the support of the Poisson distribution
}
