% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{support}
\alias{support}
\title{Return the support of a distribution}
\usage{
support(d, drop = TRUE)
}
\arguments{
\item{d}{A probability distribution object such as those created by
a call to \code{\link[=Bernoulli]{Bernoulli()}}, \code{\link[=Beta]{Beta()}}, or \code{\link[=Binomial]{Binomial()}}.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}
}
\value{
A vector with two elements indicating the range of the support.
}
\description{
Return the support of a distribution
}
