% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tukey.R
\name{Tukey}
\alias{Tukey}
\title{Create a Tukey distribution}
\usage{
Tukey(nmeans, df, nranges)
}
\arguments{
\item{nmeans}{TODO}

\item{df}{TODO}

\item{nranges}{TODO}
}
\value{
A \code{Tukey} object.
}
\description{
Tukey's studentized range distribution, used mostly for Tukey's
honestly significant differences test in ANOVA
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a Tukey random variable with
success probability \code{p} = \eqn{p}.

TODO: multiple parameterizations BLEH

\strong{Support}: TODO

\strong{Mean}: TODO

\strong{Variance}: TODO

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:

TODO
}
\examples{

X <- Tukey(4L, 16L, 2L)
X

cdf(X, 4)
quantile(X, 0.7)

}
\seealso{
Other continuous distributions: \code{\link{Beta}},
  \code{\link{Cauchy}}, \code{\link{ChiSquare}},
  \code{\link{Exponential}}, \code{\link{FisherF}},
  \code{\link{Gamma}}, \code{\link{LogNormal}},
  \code{\link{Logistic}}, \code{\link{Normal}},
  \code{\link{StudentsT}}, \code{\link{Uniform}},
  \code{\link{Weibull}}
}
\concept{continuous distributions}
