% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tukey.R
\name{Tukey}
\alias{Tukey}
\title{Create a Tukey distribution}
\usage{
Tukey(nmeans = numeric(), df = numeric(), nranges = numeric())
}
\arguments{
\item{nmeans}{Sample size for each range.}

\item{df}{Degrees of freedom.}

\item{nranges}{Number of groups being compared.}
}
\value{
A \code{Tukey} object.
}
\description{
Tukey's studentized range distribution, used for Tukey's
honestly significant differences test in ANOVA.
}
\details{
We recommend reading this documentation on
\url{https://zeileis.github.io/distributions3/}, where the math
will render with additional detail and much greater clarity.

\strong{Support}: \eqn{R^+}, the set of positive real numbers.

Other properties of Tukey's Studentized Range Distribution
are omitted, largely because the distribution is not fun
to work with.
}
\examples{

set.seed(27)

X <- Tukey(4L, 16L, 2L)
X

cdf(X, 4)
quantile(X, 0.7)
}
\seealso{
Other continuous distributions:
\code{\link[=Beta]{Beta()}},
\code{\link[=Cauchy]{Cauchy()}},
\code{\link[=ChiSquare]{ChiSquare()}},
\code{\link[=Erlang]{Erlang()}},
\code{\link[=Exponential]{Exponential()}},
\code{\link[=FisherF]{FisherF()}},
\code{\link[=Frechet]{Frechet()}},
\code{\link[=GEV]{GEV()}},
\code{\link[=GP]{GP()}},
\code{\link[=Gamma]{Gamma()}},
\code{\link[=Gumbel]{Gumbel()}},
\code{\link[=LogNormal]{LogNormal()}},
\code{\link[=Logistic]{Logistic()}},
\code{\link[=Normal]{Normal()}},
\code{\link[=RevWeibull]{RevWeibull()}},
\code{\link[=StudentsT]{StudentsT()}},
\code{\link[=Uniform]{Uniform()}},
\code{\link[=Weibull]{Weibull()}}
}
\concept{continuous distributions}
