% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.R
\name{uniform}
\alias{uniform}
\title{Create a continuous uniform distribution}
\usage{
uniform(a = 0, b = 1)
}
\arguments{
\item{a}{The a parameter. \code{a} can be any value in the set of real
numbers. Defaults to \code{0}.}

\item{b}{The a parameter. \code{b} can be any value in the set of real
numbers. It should be strictly bigger than \code{a}, but if is not, the
order of the parameters is inverted. Defaults to \code{1}.}
}
\value{
A \code{uniform} object.
}
\description{
Create a continuous uniform distribution
}
\examples{

b <- uniform(1, 2)
b

random(b, 10)
pdf(b, 0.7)
log_pdf(b, 0.7)
cdf(b, 0.7)
quantile(b, 0.7)

cdf(b, quantile(b, 0.7))
quantile(b, cdf(b, 0.7))

}
\seealso{
Other continuous distributions: \code{\link{Beta}},
  \code{\link{Cauchy}}, \code{\link{chi_square}},
  \code{\link{exponential}}, \code{\link{fisher_f}},
  \code{\link{gamma}}, \code{\link{normal}},
  \code{\link{students_t}}
}
\concept{continuous distributions}
