% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weibull.R
\name{Weibull}
\alias{Weibull}
\title{Create a Weibull distribution}
\usage{
Weibull(shape, scale)
}
\arguments{
\item{shape}{TODO}

\item{scale}{TODO}
}
\value{
A \code{Weibull} object.
}
\description{
Create a Weibull distribution
}
\details{
We recommend reading this documentation on
\url{https://alexpghayes.github.io/distributions}, where the math
will render with additional detail and much greater clarity.

In the following, let \eqn{X} be a Weibull random variable with
success probability \code{p} = \eqn{p}.

\strong{Support}: TODO

\strong{Mean}: TODO

\strong{Variance}: TODO

\strong{Probability density function (p.d.f)}:

TODO

\strong{Cumulative distribution function (c.d.f)}:

TODO

\strong{Moment generating function (m.g.f)}:

TODO
}
\examples{

X <- Weibull(0.3, 2)
X

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

}
\seealso{
Other continuous distributions: \code{\link{Beta}},
  \code{\link{Cauchy}}, \code{\link{ChiSquare}},
  \code{\link{Exponential}}, \code{\link{FisherF}},
  \code{\link{Gamma}}, \code{\link{LogNormal}},
  \code{\link{Logistic}}, \code{\link{Normal}},
  \code{\link{StudentsT}}, \code{\link{Tukey}},
  \code{\link{Uniform}}
}
\concept{continuous distributions}
